### 2019-04-17 #295 @mem48
Submission: opentripplanner
Submitting Author: Malcolm Morgan (@mem48)
Repository:   https://github.com/ITSLeeds/opentripplanner
Version submitted:   0.1.0
Version accepted: TBD

---

```
Package: opentripplanner
Title: OpenTripPlanner for R
Version: 0.1.0.0000
    person("Marcus", "Young",   email = "M.A.Young@soton.ac.uk", role = c("aut"),
    person("Malcolm", "Morgan", email = "m.morgan1@leeds.ac.uk", role = c("aut","cre"),
    person("Robin", "Lovelace", email = "rob00x@gmail.com",      role = "aut",
    person("Layik", "Hama",     email = "layik.hama@gmail.com",  role = "ctb",
    )
Maintainer: Malcolm Morgan <m.morgan1@leeds.ac.uk>
Description: An R interface for OpenTripPlanner <http://www.opentripplanner.org/>.
    OpenTripPlanner (OTP) is an open source platform for multi-modal and multi-agency
    journey planning written in Java. OTP provides a REST API which OpenTripPlanner
    for R uses to connect to OTP.
Language: EN-GB
License: file LICENSE
    checkmate,
    httr,
    sf,
    googlePolylines,
    geodist,
    pbapply
    covr
```

Open Trip Planner is a multimodal route planning tool written in Java, the package allows R users to access the functionality of Open Trip Planner running locally or on a server. So, the package performs the role of data retrieval of geospatial data.

Transport planners and people engaging in transport-related research. Route planning is the basis of many different transport related research questions such as traffic management, accessibility measurement, and infrastructure planning. The package is being actively developed to support research at the University of Leeds Institute for Transport Studies.
There are many packages that enable route planning in R (e.g. osrmr & dodgr) However there are no packages on CRAN that connect R to the Open Trip Planner. We are aware of one package on GitHub (https://github.com/datasciencecampus/access-to-services/tree/develop/propeR ) which relates to connecting R and OTP although its purpose is more specialised.
A key benefit of OTP over other route planning tools is its integration of public transport and other modes of travel (e.g. bike hire) that are not supported by other tools such as osrm.
