### 2019-04-17 57bd7e4
bumb version number
for ropensci this is first full stable version
