+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: opentripplanner
 Title: OpenTripPlanner for R
-Version: 0.0.2.0000
+Version: 0.1.0.0000
 Authors@R: c(
     person("Marcus", "Young",   email = "M.A.Young@soton.ac.uk", role = c("aut"),
       comment = c(ORCID = "0000-0003-4627-1116")),
