+++ extras/build_package.R
@@ -1,12 +1,0 @@
-# Build Package Script
-
-codemetar::give_opinions()
-styler::style_pkg()
-goodpractice::gp()
-spelling::spell_check_package()
-
-devtools::document()
-devtools::build()
-devtools::check()
-#codemetar::write_codemeta()
-#knitr::knit("README.Rmd")
