### 2019-04-20 a7c3601
test remotes fore geodist
