+++ DESCRIPTION
@@ -31,6 +31,7 @@ Imports:
     jsonlite,
     pbapply,
     sf
+Remotes: hypertidy/geodist    
 RoxygenNote: 6.1.1
 Suggests: 
     covr,
