+++ appveyor.yml
@@ -43,3 +43,10 @@ artifacts:
 
   - path: '\*_*.zip'
     name: Bits
+
+environment:
+  NOT_CRAN: true
+  # env vars that may need to be set, at least temporarily, from time to time
+  # see https://github.com/krlmlr/r-appveyor#readme for details
+  # USE_RTOOLS: true
+  # R_REMOTES_STANDALONE: true
