+++ R/otp-config.R
@@ -81,100 +81,182 @@ otp_list_clean <- function(.data, fun = function(x) {
 #' @export
 
 otp_validate_config <- function(config, type = attributes(config)$config_type) {
-  # checkmate::assert_list(config, any.missing = FALSE, names = "unique", null.ok = FALSEALSE)
-  checkmate::assert_subset(type, choices = c("otp", "build", "router"), empty.ok = FALSE)
+
+  checkmate::assert_subset(type, choices = c("otp", "build", "router"),
+                           empty.ok = FALSE)
 
   if (type == "router") {
     # Logical
-    checkmate::assert_logical(config$routingDefaults$allowBikeRental, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$arriveBy, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$batch, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$compactLegsByReversedSearch, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$disableAlertFiltering, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$disableRemainingWeightHeuristic, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$driveOnRight, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$geoidElevation, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$ignoreRealtimeUpdates, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$kissAndRide, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$longDistance, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$maxTransfers, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$onlyTransitTrips, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$parkAndRide, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$reverseOptimizeOnTheFly, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$reverseOptimizing, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$showIntermediateStops, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$softPreTransitLimiting, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$softWalkLimiting, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$useBikeRentalAvailabilityInformation, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$useRequestedDateTimeInMaxHours, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$useTraffic, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$walkingBike, len = 1, null.ok = TRUE)
-    checkmate::assert_logical(config$routingDefaults$wheelchairAccessible, len = 1, null.ok = TRUE)
+    checkmate::assert_logical(config$routingDefaults$allowBikeRental,
+                              len = 1, null.ok = TRUE)
+    checkmate::assert_logical(config$routingDefaults$arriveBy,
[…]
