+++ tests/testthat/test_03_with_OTP.R
@@ -21,10 +21,16 @@ test_that("download example data", {
   unzip(file.path(path_data, "isle-of-wight-demo.zip"),
     exdir = file.path(path_data, "graphs", "default")
   )
-  expect_true(file.exists(file.path(path_data, "graphs", "default", "isle-of-wight.osm.pbf")))
-  expect_true(file.exists(file.path(path_data, "graphs", "default", "iow-gtfs.zip")))
-  expect_true(file.exists(file.path(path_data, "graphs", "default", "IOW_DEM.tif")))
-  expect_true(file.exists(file.path(path_data, "graphs", "default", "router-config.json")))
+  expect_true(file.exists(file.path(
+    path_data, "graphs", "default", "isle-of-wight.osm.pbf")))
+  expect_true(file.exists(file.path(
+    path_data, "graphs", "default", "iow-rail-gtfs.zip")))
+  expect_true(file.exists(file.path(
+    path_data, "graphs", "default", "iow-bus-gtfs.zip")))
+  expect_true(file.exists(file.path(
+    path_data, "graphs", "default", "IOW_DEM.tif")))
+  expect_true(file.exists(file.path(
+    path_data, "graphs", "default", "router-config.json")))
   unlink(file.path(path_data, "isle-of-wight-demo.zip"))
   expect_true(!file.exists(file.path(path_data, "isle-of-wight-demo.zip")))
 })
