### 2019-05-08 6ffd204
fix sapply
