### 2019-05-09 #295 @geanders
That seems plausible, @Robinlovelace. I'm using MacOS Mojave 10.14.4.
