### 2019-05-09 8b0615b
Update otp_stop() for linux
