### 2019-05-13 478ecf0
Revert "Close #24"
This reverts commit fa7cbf102bf18aa4677dac4ce68f9df6013f4947.
