+++ tests/testthat/test_03_with_OTP.R
@@ -37,8 +37,7 @@ test_that("download otp", {
   expect_true(file.exists(file.path(path_otp)))
 })
 
-download.file("https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/centroids.gpkg", "centroids.gpkg")
-lsoa <- sf::read_sf("centroids.gpkg")
+lsoa <- sf::read_sf("https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/centroids.gpkg")
 test_that("can get lsoa points", {
   expect_is(lsoa, "sf")
   expect_true(nrow(lsoa) == 89)
