+++ tests/testthat/test_03_with_OTP.R
@@ -38,7 +38,10 @@ test_that("download example data", {
 path_otp <- file.path(path_data, "otp.jar")
 test_that("download otp", {
   url_otp <- "https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-shaded.jar"
-  download.file(url = url_otp, destfile = file.path(path_data, "otp.jar"), mode = "wb", quiet = TRUE)
+  download.file(url = url_otp,
+                destfile = file.path(path_data, "otp.jar"),
+                mode = "wb",
+                quiet = TRUE)
   expect_true(file.exists(file.path(path_otp)))
 })
 
@@ -54,7 +57,10 @@ context("Test the otp_build_graph function")
 test_that("We can build an otp graph", {
   skip_no_otp()
   log <- otp_build_graph(otp = path_otp, dir = path_data)
-  expect_true(file.exists(file.path(path_data, "graphs", "default", "Graph.obj")))
+  expect_true(file.exists(file.path(path_data,
+                                    "graphs",
+                                    "default",
+                                    "Graph.obj")))
 })
 
 context("Test the otp_setup function")
@@ -77,7 +83,8 @@ test_that("object returned when check is TRUE and router exists", {
 
 test_that("correct message when check is TRUE and router exists", {
   skip_no_otp()
-  expect_message(otp_connect(), "Router http://localhost:8080/otp/routers/default exists")
+  expect_message(otp_connect(),
+                 "Router http://localhost:8080/otp/routers/default exists")
 })
 
 test_that("correct error when check is TRUE and router does not exist", {
@@ -170,7 +177,8 @@ context("Test the otp_geocode function")
 
 test_that("basic geocode", {
   skip_no_otp()
-  stations <- otp_geocode(otpcon = otpcon, query = "station")
+  stations <- otp_geocode(otpcon = otpcon,
+                          query = "station")
   expect_is(stations, "sf")
   expect_true(nrow(stations) == 10)
 })
