### 2019-05-18 #295 @mem48
@geanders I've been chipping away at the to-do list. I think they are all done now, with a few exceptions

- There are a few 80+ character lines left where long character strings such as URLs can't easily be split
- I haven't changed the I_have_OTP for now as NOT_CRAN is similar but not exactly the same. I'm open to advice on this.

I think the otp_stop test is fixed but I don't have a mac so perhaps you or @layik could confirm?

Thanks for all your help.
