### 2019-05-22 #295 @layik
@mem48 I stumbled upon this by chance. GitHub might just be too good or I am not getting proper feeds. I will try it when I can. Just RE

> There are a few 80+ character lines left where long character strings such as URLs can't easily be split

I will see if I can split them, R's `paste0` is my friend when it comes to that, you might as well use part of those long URLs if not, it will satisfy roxygen or whatever package it is.
