### 2019-05-27 ropensci/opentripplanner#28 @kimnewzealand
OTP / Java set up issue on Windows
Hi there opentripplanner team

I am going through the installation and set up for the package review. I have updated Java to version 8 and installed and attached the R package opentripplanner, as per the vignettes. I am raising this is as an issue as I am unable to progress on the review at this point.

I initially suspected there an issue with the OTP connection as the check for OTP did not return TRUE. (I did not change the .Renviron file at that point, as suggested in the vignette, see next).
` Sys.getenv("I_have_OTP")`
>[1] ""

The build of a graph object also returned an error message:

`log <- otp_build_graph(otp = path_otp, dir = path_data) `
> Basic checks completed, building graph, this may take a few minutes
> Failed to build graph with message:
> Error occurred during initialization of VMCould not reserve enough space for 2097152KB object heap
> Warning message:
> In system(text, intern = TRUE) :
> running command 'java -Xmx2G -jar > >"C:\Users\Home\AppData\Local\Temp\RtmpoNFxd9/OTP/otp.jar" --build >"C:\Users\Home\AppData\Local\Temp\RtmpoNFxd9/OTP/graphs/default"' had status 1

I did some Googling and tried this fix on the environment variables in the control panel:
System Variables->New:
Variable name: _JAVA_OPTIONS
Variable value: -Xmx2G

Edit Environment Variable path:
Variable name: Path
Variable value: ;C:\Program Files\Java\jre6\bin;F:\JDK\bin;

I now see a successful OTP connection,

` Sys.getenv("I_have_OTP")`
> [1] "TRUE"

I then retried the graph object  and perhaps the new variables are good as it looks like it has gone past the initialisation,  but there appears to be a Java version issue now. ? I am not familiar with Java so I was wondering if you could advise on what to try next?

> # Run the OTP and Build a graph object
` log <- otp_build_graph(otp = path_otp, dir = path_data) `
> Error in otp_checks(otp = otp, dir = dir, router = router, graph = FALSE) :

> In addition: Warning messages:
>1: In system2("java", "-version", stdout = TRUE, stderr = TRUE) :
>  running command '"java" -version' had status 1
>2: In otp_checks(otp = otp, dir = dir, router = router, graph = FALSE) :
>  NAs introduced by coercion
>3: In otp_checks(otp = otp, dir = dir, router = router, graph = FALSE) :

Also see below the session information:

` sessionInfo()`
R version 3.5.1 (2018-07-02)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows >= 8 x64 (build 9200)

Matrix products: default

locale:
[1] LC_COLLATE=English_New Zealand.1252  LC_CTYPE=English_New Zealand.1252
[3] LC_MONETARY=English_New Zealand.1252 LC_NUMERIC=C
[5] LC_TIME=English_New Zealand.1252

attached base packages:

other attached packages:
[1] opentripplanner_0.1.0.0000

 [1] compiler_3.5.1   backports_1.1.4  clisymbols_1.2.0 tools_3.5.1      fs_1.2.6
 [6] glue_1.3.1       rstudioapi_0.9.0 yaml_2.2.0       crayon_1.3.4     Rcpp_1.0.1
[11] usethis_1.5.0    checkmate_1.9.3  rlang_0.3.4
