### 2019-05-27 #295 @kimnewzealand
I have raised an issue (see here https://github.com/ITSLeeds/opentripplanner/issues/28)
