### 2019-05-28 ropensci/opentripplanner#28 @mem48
Thanks for the detailed responce I'll try to reprodduce the issue, and get back to you asap
