### 2019-05-28 985fc42
add numItineraries to the API query
