### 2019-05-29 ropensci/opentripplanner#28 @kimnewzealand
thanks @mem48

I am trying on another laptop now so I will update thread soon
