### 2019-05-29 ropensci/opentripplanner#28 @kimnewzealand
As an update, I have uninstalled and reinstalled Java version 8 on my machine and installed the package on another laptop with more memory.

I will ignore Sys.getenv("I_have_OTP") for now as you suggest.

I tried to run this command and received an error:
` opentripplanner:::otp_checks()`
> Error in opentripplanner:::otp_checks() :
>  Assertion on 'dir' failed: No directory provided.

1. Despite the error with `opentripplanner:::otp_checks()`,  I had success in building the graph by adding memory =1 on both laptops, see below message:

> # Run the OTP and Build a graph object
` log <- otp_build_graph(otp = path_otp, dir = path_data, memory = 1) `
> Basic checks completed, building graph, this may take a few minutes
> Graph built

2. I removed the path and variables as this didn't appear to be necessary on the other laptop. I suspect that the Java upgrade didn't work the first time round. Potentially Java needed a full uninstall first.

`C:\Users\Home>java -version`
> java version "1.8.0_211"
> Java(TM) SE Runtime Environment (build 1.8.0_211-b12)
> Java HotSpot(TM) Client VM (build 25.211-b12, mixed mode, sharing)

Any further thoughts? Else this can probably be closed
