### 2019-05-29 ropensci/opentripplanner#30 @kimnewzealand
Set up an OTP instance taking too long
I have successfully created a graph object but it seems like there is a potential issue with setting up the OTP instance. See below error message.

` log <- otp_build_graph(otp = path_otp, dir = path_data, memory = 1) `
> Basic checks completed, building graph, this may take a few minutes
> Graph built

` otp_setup(otp = path_otp, dir = path_data)`
> 2019-05-30 11:05:41 OTP is loading and may take a while to be useable
> 2019-05-30 11:11:05 OTP is taking an unusually long time to load, releasing R to your control

## Trying the other alternative ports example
` otp_setup(otp = path_otp, dir = path_data, port = 8801, securePort = 8802)`
> 2019-05-30 11:11:06 OTP is loading and may take a while to be useable
> 2019-05-30 11:16:29 OTP is taking an unusually long time to load, releasing R to your control
