### 2019-05-29 #295 @kimnewzealand
Thanks for the help with the issue #28.  That can probably be closed?

I have raised another issue ( see here https://github.com/ITSLeeds/opentripplanner/issues/30)
