### 2019-05-30 ropensci/opentripplanner#30 @kimnewzealand
O this one does not appear to be an issue. I tried the memory=1 trick again and it worked so I will close

` otp_setup(otp = path_otp, dir = path_data,memory=1)`
> 2019-05-30 12:02:35 OTP is loading and may take a while to be useable
> Router http://localhost:8080/otp/routers/default exists
> 2019-05-30 12:05:47 OTP is ready to use Go to localhost:8080 in your browser to view the OTP
