### 2019-05-30 ropensci/opentripplanner#28 @mem48
Seems like its fixed, the otp_ckecks requires several arguments to work, sorry I did not make that clear before. So the error is expected. I think for now I will update the vignettes to mention memory requirements and perhaps some more tips on installing Java.
