### 2019-05-30 2e70f14
Add hardware specs
