### 2019-05-30 #295 @mem48
On the "OTP is taking an unusually long time to load, releasing R to your control" this might a bit confusing.

Because loading OTP takes a while, for my work I have large setups that can take several days to start up. I wrote the code so that R will wait for no more than 5 minutes before giving up. But that does not mean that OTP has stopped, in fact, OTP continues without R.

On refelction this may be confusing for users, os perhaps I should change the behaviour to be the same as otp_build_graph, where R will wait until OTP is ready?
