### 2019-05-30 f71c8fc
Fix @robinlovelace bug
Fix for linux working introduced a bug, nof fixed also cleaned code and added check for small graphs
