### 2019-06-05 #295 @kimnewzealand
Thanks again for asking me to perform this review.

I have regular user perspective of someone who may be using this package personally or in a consultancy.

I spent the bulk of the time working though the Documentation, thinking of points where a new user may face issues using the package.

I ran the tests check but I did not use a larger data set as I also only have access to an 8GB RAM laptop. Therefore I did not test for other use cases or edge cases.

Let me know if you have any further questions.

- [x ] **Installation instructions:** for the development version of package and any non-standard dependencies in README

##### For packages co-submitting to JOSS

- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)

- [x] **Performance:** Any performance claims of the software been confirmed. *Only with 8MG Memory and given Isle of Wight sample dataset*
   and a reasonable range of inputs and conditions. All tests pass on the local machine. ** see results below**

---

This is a great package providing an open source alternative for routing, to calculate routes using OTP from OSM, combine with other data and also create visualisations with this data - all within R.

As a user, once you get to `Launch OTP and load the graph` it is really exciting to see this part working visually with the OTP connection.

####  Package Review Checklist - Documentation

Here are some further notes on the Documentation checklist above.

**Examples**

I did not check off the Examples in the checklist above. I couldn't see a function nor find an example for it in the vignettes for this R script  `opentripplanner-package.R`, is  the R script needed?

The R help files are also comprehensive however the Examples, may need a review and update. They all include `Not run` comments which can probably be removed.

I appreciate that there is a sequential nature to most of the functions, perhaps use the vignette examples here or just provide a link to the getting started vignette?

**Vignettes**

Initially it took me half day (in total) to get setup due my laptop's memory issues with the Java and OTP tool prerequisites. However the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html) has been updated with hardware considerations for the first time user based on issues raised.

The vignettes are fairly long and comprehensive which I think is needed for a first time user, given the initial set up required.

A comment regarding the target audience is that if the user is also new to R, then they may not be familiar with the different data formats and how to read those into R. From what I understand the `opentripplanner` package will read in any data format as long as it is downloaded into the expected folder structure.

**Suggested updates**

I have a couple of documentation suggestions:

- I suggest moving `Getting Data for OTP` section from the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html) and the Isle of Wight data link to the [opentripplanner: getting started](https://itsleeds.github.io/opentripplanner/articles/opentripplanner.html) vignette. It feels like this step follows the installation of the R package, but is related to `Building an OTP Graph` and the expectations around folder structure.
- Rename the the [opentripplanner: getting started](https://itsleeds.github.io/opentripplanner/articles/opentripplanner.html) vignette to `getting started`
- Add the getting started to the Articles drop down menu
- Depending on rOpenSci's current view on Datacamp, perhaps remove the tutorial  references
- Add a reference hyperlink for the .pbf format
- Add a reference hyperlink for the GeoTIFF format
- In this line `Note that OTP only really needs the road network from the OSM. So for large areas, you may wish to remove unneeded data from the file using osmconvert and osmfilter and` remove the `and` and explain where (OSM or R?) you can remove this unneeded data.
- Move this section from  `getting started` to `prerequisites` as it feels like it part of the initial set up, and the user can work through this set up once before then installing or loading the R package.
> Before you can use the package, you should set-up the OpenTripPlanner Java program. You will need to download the latest version of OTP from https://repo1.maven.org/maven2/org/opentripplanner/otp/ , at time of writing the latest version of OTP is version 1.3 so you would want to get https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-shaded.jar

- Include a mention of the memory parameter or include another line of code in case the user encounters memory issues in this step:
`log <- otp_build_graph(otp = path_otp, dir = path_data, memory = 1) `
- There are few places ion the README and vignettes where the `.` is missing at the end of the sentences

**Community guidelines**

I did not check off the Community guidelines in the checklist above. Although there is a Contributor code of conduct there is no markdown
guide for how to contribute.

#### Package Review Checklist - For packages co-submitting to JOSS

I did not check off the References in the checklist above as it is blank. Should this be populated with the Isle of Wight data source?

Suggested update:

- Update the wording in the link to `SF data.frame ` to `sf data frame` as data.frame is the function whereas I believe the object is called a data frame.
- Update `A full explanation is provided in the package vignettes` link to https://itsleeds.github.io/opentripplanner/articles/

####  Package Review Checklist - Functionality

I checked all items in the Functionality checklist, however I have made some suggested updates below.

**Suggested updates**

I have a couple of functionality suggestions regarding messages:

- Ensure consistency across success messages, and include progress bars for example:

The success message is simple when running `otp_build`.

> Basic checks completed, building graph, this may take a few minutes
> Graph built

 but the success message for `otp_setup` has a different format

 > otp_setup(otp = path_otp, dir = path_data, memory=1)
2019-06-05 14:47:22 OTP is loading and may take a while to be useable
Router http://localhost:8080/otp/routers/default exists
2019-06-05 14:50:07 OTP is ready to use Go to localhost:8080 in your browser to view the OTP

and the `otp_plan` has a nice progress bar but it is missing a message description of what's happening

>|++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 14s

- The error message for `ot_build_graph` low memory is complicated, could this include a recommendation to add the memory parameter?

`log <- otp_build_graph(otp = path_otp, dir = path_data )`
>Basic checks completed, building graph, this may take a few minutes
Failed to build graph with message:
Error occurred during initialization of VMCould not reserve enough space for 2097152KB object heap
In system(text, intern = TRUE) :
  running command 'java -Xmx2G -jar "C:\Users\Home\AppData\Local\Temp\RtmpSYvN4O/OTP/otp.jar" --build "C:\Users\Home\AppData\Local\Temp\RtmpSYvN4O/OTP/graphs/default"' had status 1
>

- Perhaps don't include the timer cut-off of 5 minutes  with this message, or include a parameter for this cut-off. The message is not useful as a user doesn't know what action to take:
`otp_setup(otp = path_otp, dir = path_data)`

> 2019-05-30 11:05:41 OTP is loading and may take a while to be useable
> 2019-05-30 11:11:05 OTP is taking an unusually long time to load, releasing R to your control

**Results of package checks and tests**

-- R CMD check results ---------------------------------------------- opentripplanner 0.1.0.0000 ----
Duration: 6m 10.2s

> checking for unstated dependencies in examples ... OK
   WARNING
  'qpdf' is needed for checks on size reduction of PDFs

> checking top-level files ... NOTE
  Non-standard file/directory found at top level:
    'paper.bib'

> checking R code for possible problems ... NOTE
  Warning: <anonymous>: ... may be used in an incorrect context

  otp_get_results: ... may be used in an incorrect context

0 errors √ | 1 warning x | 2 notes x

> devtools::test()
Loading opentripplanner
Testing opentripplanner
√ | OK F W S | Context
   |                                                  | 0 % ~calculating
   |                                                  | 0 % ~calculating
   |                                                  | 0 % ~calculating
   |                                                  | 0 % ~calculating
√ | 29   3   | Test function without an OTP connection [21.2 s]
test_01_without_OTP.R:55: warning: otp_plan input validation
repeating fromPlace to match length of toPlace

test_01_without_OTP.R:61: warning: otp_plan input validation
repeating toPlace to match length of fromPlace

test_01_without_OTP.R:68: warning: otp_plan input validation
repeating toPlace to match length of fromPlace
√ |  0       | Test internal functions
√ |  2       | Test internal functions from otp-config.R
√ | 19       | Test internal functions from otp-plan.R [0.7 s]
√ | 10   3   | Download required files [56.4 s]
test_03_with_OTP.R:13: warning: download example data
'C:\Users\Home\AppData\Local\Temp\RtmpuSSdY2\otp' already exists

test_03_with_OTP.R:14: warning: download example data
'C:\Users\Home\AppData\Local\Temp\RtmpuSSdY2\otp\graphs' already exists

test_03_with_OTP.R:15: warning: download example data
'C:\Users\Home\AppData\Local\Temp\RtmpuSSdY2\otp\graphs\default' already exists
√ |  0     1 | Test the otp_build_graph function
test_03_with_OTP.R:60: skip: We can build an otp graph
Not running full test.
√ |  0     1 | Test the otp_setup function
test_03_with_OTP.R:71: skip: We can startup OTP
Not running full test.
√ |  0     3 | Test the otp_connect function [0.1 s]
test_03_with_OTP.R:81: skip: object returned when check is TRUE and router exists
Not running full test.

test_03_with_OTP.R:87: skip: correct message when check is TRUE and router exists
Not running full test.

test_03_with_OTP.R:93: skip: correct error when check is TRUE and router does not exist
Not running full test.
√ |  0     5 | Test the otp_plan function [0.1 s]
test_03_with_OTP.R:105: skip: basic routing
Not running full test.

test_03_with_OTP.R:116: skip: transit routing
Not running full test.

test_03_with_OTP.R:129: skip: no geometry routing
Not running full test.

test_03_with_OTP.R:140: skip: full elevation routing
Not running full test.

test_03_with_OTP.R:152: skip: batch routing
Not running full test.
√ |  0     2 | Test the otp_isochone function
test_03_with_OTP.R:166: skip: basic isochrone
Not running full test.

test_03_with_OTP.R:180: skip: nonsence isochrone
Not running full test.
√ |  0     5 | Test the otp_geocode function [0.1 s]
test_03_with_OTP.R:189: skip: basic geocode
Not running full test.

test_03_with_OTP.R:198: skip: geocode coords
Not running full test.

test_03_with_OTP.R:206: skip: geocode both
Not running full test.

test_03_with_OTP.R:214: skip: geocode nonsence
Not running full test.

test_03_with_OTP.R:222: skip: otp_stop
Not running full test.
   |++++++++++++++++++++++++++++++++++++++++++++++++++| 100% elapsed = 01s
√ |  2       | Test the otp_plan function [1.1 s]
√ |  2       | Test the otp_isochone function [19.2 s]
√ |  2       | Test the otp_geocode function [0.5 s]

== Results =====================================================================

OK:       66

Woot!
1: closing unused connection 4 (<-LAPTOP-222VVQ3R:11467)
2: closing unused connection 3 (<-LAPTOP-222VVQ3R:11467)
