### 2019-06-05 #295 @mem48
Thanks for the review, lots to get stuck into. Most of it I completly agree with so I'll get started. There are a couple of point that I think will be technically very difficult (around error messages etc) I'll see what I can do and get back to you.
