### 2019-06-05 #295 @mem48
@geanders thanks for the heads up, I'll get started on some of the basics like typos and links then wait for the second review for the more complex issues.
