+++ vignettes/prerequisites.Rmd
@@ -107,40 +107,21 @@ If you have never used R before there are lots of free tutorials to get you star
 
 ### To use OTP
 
-Download and install Java. You will need the correct version for your operating system. If possible the 64 Bit version of Java is preferable, especially if you want to use OTP over large areas.
+Download and install Java 8. You will need the correct version for your operating system. If possible the 64 Bit version of Java is preferable, especially if you want to use OTP over large areas.
 
 **Note** OTP does not work with later versions of Java such as Java 11.
 
 To get Java:
 
-* **Windows/Mac** https://www.java.com/en/download/ This link defaults to the 32 Bit version.
+* **Windows/Mac** https://www.java.com/en/download/ This link defaults to the 32 Bit version, so also check https://java.com/en/download/manual.jsp.
 
 * **Linux**  we recommend instructions at [StackOverflow](https://askubuntu.com/questions/740757/switch-between-multiple-java-versions).
 
+
+Before you can use the package, you should set-up the OpenTripPlanner Java program.
+You will need to download the latest version of OTP from https://repo1.maven.org/maven2/org/opentripplanner/otp/ , at time of writing the latest version of OTP is version 1.3 so you would want to get https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-shaded.jar 
+
 You can also compile the latest version of OTP from source, as outlined at [docs.opentripplanner.org/](http://docs.opentripplanner.org/en/latest/Getting-OTP/).
-
-## Getting Data for OTP
-
-To use OTP on your local computer you will need data about your location of interest, such as where the roads are, what is the public transport timetable etc.
-
-### Road Map Data
-
-OTP uses road maps from the [Open Street Map (OSM)](https://www.openstreetmap.org) in the .pbf format. OSM is a free map that anybody can edit. You can download all or part of the OSM from [Geofabrik](https://download.geofabrik.de/).
-
-Note that OTP only really needs the road network from the OSM. So for large areas, you may wish to remove unneeded data from the file using [osmconvert](https://wiki.openstreetmap.org/wiki/Osmconvert) and [osmfilter](https://wiki.openstreetmap.org/wiki/Osmfilter) and 
-
-### Public Transport Timetables
-
-OTP can use public transport timetable data in the [GTFS format](https://developers.google.com/transit/gtfs/). You can find GTFS data for many regions at  [transitland](https://transit.land/).
-
-### Elevation Data
-
-You can add terrain information to your routes, especially useful for walking and cycling using GeoTIFF images. You can find worldwide elevation data from [NASA](https://www2.jpl.nasa.gov/srtm/).
-
-**Warning** It is common for GeoTIFF to have a no data value often the maximum possible value. OTP can misinterpret this as an elevation value. So set your no data values in your elevation data to something more plausible like 0.
-
-Note that OTP does not support all types of GeoTIFF compression so you may have to change the compression type of the image if you are experiencing problems. 
-
 
 ## Next Steps
 
