+++ vignettes/opentripplanner.Rmd
@@ -14,7 +14,7 @@ vignette: >
 This vignette is an introduction to OpenTripPlanner (OTP) - an open-source and cross-platform multi-modal route planner written in JAVA. It uses imported Open Street Map (OSM) data for routing on the street and path network and supports
 multi-agency public transport routing through imported  [General Transit Feed Specification](https://developers.google.com/transit/gtfs/) (GTFS) feeds. It can also apply a digital elevation model to the OSM street network, allowing, for example, cycle-friendly routes to be requested. OTP has a web front-end that can be used by end-users and a sophisticated routing API.
 
-OTP works worldwide as long as there is OSM map coverage in the area. Support for transit timetables and elevation calculations are dependant on the available data. For more information on getting data see the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html)
+OTP works worldwide as long as there is OSM map coverage in the area. Support for transit timetables and elevation calculations are dependant on the available data. For more information on getting data see the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html).
 
 A major advantage of running your own multi-modal route planner is the ability to carry out analysis using amended transport data. Services such as Google Maps or TransportAPI are based on current public transport schedules and the existing road network. OTP enables you to modify transit schedules and/or make changes to the underlying street network. By editing a local copy of OSM, you can model the effects of opening new roads, closing roads or imposing other restrictions. You can also look back in time. For example, you might want to examine the effect of reductions in rural bus services on the accessibility of health facilities. To do this you would need a network with bus schedules as they were in previous years.
 
@@ -27,14 +27,13 @@ You will need to have installed R, RStudio, and Java 8. For more details on the 
 The following commands will install and attach the package:
 
 ```{r, eval=FALSE}
-# install.packages("devtools")
-devtools::install_github("ITSLeeds/opentripplanner")
+remotes::install_github("ITSLeeds/opentripplanner") # Requires the remotes package
 library(opentripplanner)
 ```
 
-Before you can use the package, you should set-up the OpenTripPlanner Java program.
-You will need to download the latest version of OTP from https://repo1.maven.org/maven2/org/opentripplanner/otp/ , at time of writing the latest version of OTP is version 1.3 so you would want to get https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-shaded.jar 
+## Downloading OTP
 
+Then you will need to download the OTP, for more details see the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html).
 The following code downloads the file into a temporary folder called `OTP` you may wish to change this to keep your results after closing R:
 
 ```{r, eval=FALSE}
@@ -46,12 +45,9 @@ u <- "https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-sha
 download.file(url = u, destfile = path_otp, mode = "wb")
 ```
 
-For more information and advanced options, see http://docs.opentripplanner.org/en/latest/Getting-OTP/ 
-
-## Building an OTP Graph
+## Downloading Example Data
 
 We will build an example graph for the Isle of Wight using some example data provided for the package. A graph is what OTP uses to find routes, and must be built out of the raw data provided. First, we will download the sample data and put it in the folder structure that OTP expects.
-
 
 ```{r eval=FALSE}
 # Path to the sample data
@@ -64,6 +60,8 @@ unzip(file.path(path_data,"isle-of-wight-demo.zip"), exdir = file.path(path_data
 unlink(file.path(path_data,"isle-of-wight-demo.zip"))
 ```
 
+## Building an OTP Graph
+
[…]

[…]
