### 2019-06-07 392462b
Tidy code
