### 2019-06-07 e86130b
fix bug in tests
