### 2019-06-07 ce50bc7
Tweaks to the setup code
Use MB not GB improve wiating handelling
