+++ R/otp-setup.R
@@ -7,7 +7,7 @@
 #'
 #' @param otp A character string, path to the OTP .jar file
 #' @param dir A character string, path to a directory containing the necessary files, see details
-#' @param memory A positive integer. Amount of memory to assign to the OTP in GB, default is 2
+#' @param memory A positive integer. Amount of memory to assign to the OTP in MB, default is 2048
 #' @param router A character string for the name of the router, must match with contents of dir, default "default"
 #' @param analyst Logical, should analyst feature be built, default FALSE
 #' @return
@@ -40,18 +40,23 @@
 #' @export
 otp_build_graph <- function(otp = NULL,
                             dir = NULL,
-                            memory = 2,
+                            memory = 2048,
                             router = "default",
                             analyst = FALSE) {
 
   # Run Checks
-  otp_checks(otp = otp, dir = dir, router = router, graph = FALSE)
-  message("Basic checks completed, building graph, this may take a few minutes")
+  checkmate::assert_numeric(memory, lower = 500)
+  check <- otp_checks(otp = otp, dir = dir, router = router, graph = FALSE)
+  if(!check){
+    stop()
+  }
+  message(paste0(Sys.time(),
+                 " Basic checks completed, building graph, this may take a few minutes"))
 
   text <- paste0(
     "java -Xmx",
     memory,
-    'G -jar "',
+    'M -jar "',
     otp,
     '" --build "',
     dir,
@@ -68,10 +73,10 @@ otp_build_graph <- function(otp = NULL,
 
   # Check for errors
   if (any(grepl("ERROR", set_up, ignore.case = TRUE)) & length(set_up) < 10) {
-    message("Failed to build graph with message:")
+    message(paste0(Sys.time()," Failed to build graph with message:"))
     message(set_up)
   } else {
-    message("Graph built")
+    message(paste0(Sys.time()," Graph built"))
   }
   return(set_up)
 }
@@ -87,7 +92,7 @@ otp_build_graph <- function(otp = NULL,
 #' The function assumes you have run otp_build_graph()
 #' @param otp A character string, path to the OTP .jar file
 #' @param dir A character string, path to a directory containing the necessary files, see details
-#' @param memory A positive integer. Amount of memory to assign to the OTP in GB, default is 2
+#' @param memory A positive integer. Amount of memory to assign to the OTP in MB, default is 2048
 #' @param router A character vector for the name of the routers, must match with contents of dir, default "default"
 #' Only a single router is currently supported
 #' @param port A positive integer. Optional, default is 8080.
@@ -113,20 +118,24 @@ otp_build_graph <- function(otp = NULL,
 #' @examples
 #' \dontrun{
 #' otp_setup(otp = "C:/otp/otp.jar", dir = "C:/data")
-#' otp_setup(otp = "C:/otp/otp.jar", dir = "C:/data", memory = 5, analyst = TRUE)
+#' otp_setup(otp = "C:/otp/otp.jar", dir = "C:/data", memory = 5000, analyst = TRUE)
 #' }
 #' @export
 otp_setup <- function(otp = NULL,
                       dir = NULL,
-                      memory = 2,
+                      memory = 2048,
                       router = "default",
                       port = 8080,
                       securePort = 8081,
                       analyst = FALSE,
                       wait = TRUE) {
   # Run Checks
-  otp_checks(otp = otp, dir = dir, router = router, graph = TRUE)
-  memory <- floor(memory) # Can't have fractions of GB
[…]
