### 2019-06-07 #295 @mem48
Some comments on @kimnewzealand 's review:

> I couldn't see a function nor find an example for it in the vignettes for this R script opentripplanner-package.R, is the R script needed?

This script makes the package help file. Try ?opentripplanner in the console. Although not necessary it is recommended in r-pkgs.had.co.nz/man.html

> The R help files are also comprehensive however the Examples, may need a review and update. They all include Not run comments which can probably be removed.

I've removed some `dontrun` tags but I think in most cases the functions will not run correctly in isolation and would cause crashes or other bugs if removed. I could be wrong perhaps you can clarify why they should be removed?

> Rename the the opentripplanner: getting started vignette to getting started
> Add the getting started to the Articles drop down menu

This is an automated process for the GitHub pages, it adds a vignette with the same name as the package to the menu bar. This is the default for [pkgdown](https://pkgdown.r-lib.org/) is there a reason to now use the standard approach?

> I did not check off the References in the checklist above as it is blank. Should this be populated with the Isle of Wight data source?

References are in https://github.com/ITSLeeds/opentripplanner/blob/master/paper.bib
