### 2019-06-09 #295 @kimnewzealand
@geanders please could you confirm standard practice for the paper References to JOSS ? is it as formatted and in a separate .bib file?

> > I did not check off the References in the checklist above as it is blank. Should this be populated with the Isle of Wight data source?
>
> References are in https://github.com/ITSLeeds/opentripplanner/blob/master/paper.bib
