### 2019-06-09 #295 @kimnewzealand
I was thinking from a usability perspective  that intuitively you may not copy and run code that is tagged `Not run`. So perhaps the question should be how would a use use and follow each function's help file if they had not read the vignette? and if most functions do not run in isolation then are there a some wording or links that could be added to each help file so show the function dependencies?

> > The R help files are also comprehensive however the Examples, may need a review and update. They all include Not run comments which can probably be removed.
>
> I've removed some `dontrun` tags but I think in most cases the functions will not run correctly in isolation and would cause crashes or other bugs if removed. I could be wrong perhaps you can clarify why they should be removed?
