### 2019-06-09 #295 @kimnewzealand
> Some comments on @kimnewzealand 's review:
>
> > I couldn't see a function nor find an example for it in the vignettes for this R script opentripplanner-package.R, is the R script needed?
>
> This script makes the package help file. Try ?opentripplanner in the console. Although not necessary it is recommended in r-pkgs.had.co.nz/man.html
>
> > Rename the the opentripplanner: getting started vignette to getting started
> > Add the getting started to the Articles drop down menu
>
> This is an automated process for the GitHub pages, it adds a vignette with the same name as the package to the menu bar. This is the default for [pkgdown](https://pkgdown.r-lib.org/) is there a reason to now use the standard approach?
>

Thanks for looking into these points. The responses make sense so there is no need to make changes..
