### 2019-06-10 184dcbc
Remove unnecessary warning message
