### 2019-06-12 #295 @mem48
Thanks, I think I will wait for the second reviewer before making any further changes.

> Perhaps don't include the timer cut-off of 5 minutes with this message, or include a parameter for this cut-off. The message is not useful as a user doesn't know what action to take:

I had a look at changing this, but it causes all sorts of problems like R crashing, or just waiting forever. Which is probably why I wrote the code this way in the first place. But what I can do is increase the maximum wait time so that users are much less likely to encounter the message.
