+++ man/otp_write_config.Rd
@@ -17,9 +17,9 @@ otp_write_config(config, dir = NULL, router = "default")
 Takes a config list produced by `otp_make_config()` and saves it as json file for OTP
 }
 \examples{
-\dontrun{
+{
 conf <- otp_make_config("build")
-otp_write_config(conf, dir = "C:/otp")
+otp_write_config(conf, dir = temp.dir())
 }
 }
 \seealso{
