### 2019-06-13 581267e
stop exmaples running
due to path errors
