### 2019-08-12 ed98b2c
Start known issues vignette
