### 2019-08-12 9936088
Fix for 1.4
Fix for #32
