### 2019-08-12 319e44c
fix vignettes
