+++ vignettes/prerequisites.Rmd
@@ -4,7 +4,7 @@ author: "Malcolm Morgan"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{opentripplanner-dummies}
+  %\VignetteIndexEntry{opentripplanner-prerequisites}
   %\VignetteEngine{knitr::rmarkdown}
   %\VignetteEncoding{UTF-8}
 ---
