+++ vignettes/opentripplanner.Rmd
@@ -41,7 +41,7 @@ The following code downloads the file into a temporary folder called `OTP` you m
 dir.create(file.path(tempdir(), "OTP"))
 path_data <- file.path(tempdir(), "OTP")
 path_otp <- file.path(path_data, "otp.jar")
-u <- "https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-shaded.jar"
+u <- "https://repo1.maven.org/maven2/org/opentripplanner/otp/1.4.0/otp-1.4.0-shaded.jar"
 download.file(url = u, destfile = path_otp, mode = "wb")
 ```
 
