### 2019-08-15 46e67fd
Update multicore docs
