### 2019-08-15 23a56fc
Update known_issues.Rmd
