### 2019-09-01 9e1ff8c
fix scientific notification bug
