### 2019-10-03 ddc4463
disable public api tests
Plannerstack is down, so causing build to fail
