### 2019-10-04 1f52403
tidy vignettes
