+++ vignettes/prerequisites.Rmd
@@ -85,31 +85,17 @@ OTP is optimised for city-scale routing and performance will degrade with larger
 
 **Note** If you use a 32 Bit version of Java the maximum amount of memory that can be used by OTP is 4 GB.
 
+### Software
 
+The OpenTripPlanner for R package requires:
 
-### To Use R
+* R - [download R](https://cran.r-project.org/mirrors.html) selecting your country (or nearest available country)
+* RStudio - [download free version](https://www.rstudio.com/products/rstudio/download/) for personal use. RStudio is not essential, but is strongly reccomneded.
+* Java 8 **Note** OTP does not work with later versions of Java such as Java 11.
+* OpenTripPlanner jar file [download](https://repo1.maven.org/maven2/org/opentripplanner/otp/) or [compile from source](http://docs.opentripplanner.org/en/latest/Getting-OTP/)
 
-* Download free R software https://cran.r-project.org/mirrors.html selecting your country (or nearest available country)
-* We strongly recommend you also download and install RStudio https://www.rstudio.com/products/rstudio/download/ that has a free version for personal use.
-
-A video tutorial for installing R and RStudio 
-
-* On Mac https://www.youtube.com/watch?v=cX532N_XLIs 
-* On Windows https://www.youtube.com/watch?v=9-RrkJQQYqY 
-
-### Learning R
-
-If you have never used R before there are lots of free tutorials to get you started, such as:
-
-* https://www.datacamp.com/courses/free-introduction-to-r
-* http://www.r-tutor.com/r-introduction
-
-
-### To use OTP
 
 Download and install Java 8. You will need the correct version for your operating system. If possible the 64 Bit version of Java is preferable, especially if you want to use OTP over large areas.
-
-**Note** OTP does not work with later versions of Java such as Java 11.
 
 To get Java:
 
@@ -118,10 +104,18 @@ To get Java:
 * **Linux**  we recommend instructions at [StackOverflow](https://askubuntu.com/questions/740757/switch-between-multiple-java-versions).
 
 
-Before you can use the package, you should set-up the OpenTripPlanner Java program.
-You will need to download the latest version of OTP from https://repo1.maven.org/maven2/org/opentripplanner/otp/ , at time of writing the latest version of OTP is version 1.3 so you would want to get https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-shaded.jar 
+#### Help with R
 
-You can also compile the latest version of OTP from source, as outlined at [docs.opentripplanner.org/](http://docs.opentripplanner.org/en/latest/Getting-OTP/).
+A video tutorial for installing R and RStudio 
+
+* On Mac https://www.youtube.com/watch?v=cX532N_XLIs 
+* On Windows https://www.youtube.com/watch?v=9-RrkJQQYqY 
+
+If you have never used R before there are lots of free tutorials to get you started, such as:
+
+* https://www.datacamp.com/courses/free-introduction-to-r
+* http://www.r-tutor.com/r-introduction
+
 
 ## Next Steps
 
