### 2019-10-10 8d642e6
Reintroduce public api test and support fares
