+++ man/otp_connect.Rd
@@ -4,14 +4,16 @@
 \alias{otp_connect}
 \title{Set up and confirm a connection to an OTP instance.}
 \usage{
-otp_connect(hostname = "localhost", router = "default", port = 8080,
-  ssl = FALSE, check = TRUE)
+otp_connect(hostname = "localhost", router = "default", url = NULL,
+  port = 8080, ssl = FALSE, check = TRUE)
 }
 \arguments{
 \item{hostname}{A string, e.g. "ec2-34-217-73-26.us-west-2.compute.amazonaws.com".
 Optional, default is "localhost".}
 
 \item{router}{A string, e.g. "UK2018". Optional, default is "default".}
+
+\item{url}{If a nonstandard URL strucutre is used provide a full url, default is NULL}
 
 \item{port}{A positive integer. Optional, default is 8080.}
 
@@ -40,6 +42,9 @@ otpcon <- otp_connect(
   router = "UK2018",
   port = 8888,
   ssl = TRUE
+)
+otpcon <- otp_connect(
+  url = "https://api.digitransit.fi/routing/v1/routers/hsl"
 )
 }
 }
