+++ vignettes/opentripplanner.Rmd
@@ -116,7 +116,7 @@ If you have not used the default settings, such as a different `port` you can sp
 otpcon <- otp_connect(hostname =  "localhost", router = "default", port = 8801)
 ```
 
-You can also use the `otp_connect()` function to make a connection to OTP on a remote server, for example [digitransit](https://digitransit.fi/en/) run an OTP server for Finland. See their [documentation](https://digitransit.fi/en/developers/) for more details. Digitransit use a non-standard URL strucutre for their API, so we provide the full URL directly to `otp_connect()`.
+You can also use the `otp_connect()` function to make a connection to OTP on a remote server, for example [digitransit](https://digitransit.fi/en/) run an OTP server for Finland. See their [documentation](https://digitransit.fi/en/developers/) for more details. Digitransit use a non-standard URL strucutre for their API, so we provide the full URL directly to `otp_connect()`. See the `otp_connect()` help for more details.
 
 ```{r, eval = FALSE}
 otpcon <- otp_connect(url = "https://api.digitransit.fi/routing/v1/routers/hsl")
