+++ R/otp-setup.R
@@ -1,26 +1,32 @@
 #' Build an OTP Graph
 #'
 #' @description
-#' OTP is run in Java and requires Java commands to be typed into the command line.
-#' The function allows the parameters to be defined in R and automatically passed to Java.
-#' This function builds a OTP graph from the Open Street Map and other files.
+#' OTP is run in Java and requires Java commands to be typed into the
+#' command line. The function allows the parameters to be defined in
+#' R and automatically passed to Java. This function builds a OTP
+#' graph from the Open Street Map and other files.
 #'
 #' @param otp A character string, path to the OTP .jar file
-#' @param dir A character string, path to a directory containing the necessary files, see details
-#' @param memory A positive integer. Amount of memory to assign to the OTP in MB, default is 2048
-#' @param router A character string for the name of the router, must match with contents of dir, default "default"
+#' @param dir A character string, path to a directory containing the
+#'     necessary files, see details
+#' @param memory A positive integer. Amount of memory to assign to
+#'     the OTP in MB, default is 2048
+#' @param router A character string for the name of the router, must
+#'     match with contents of dir, default "default"
 #' @param analyst Logical, should analyst feature be built, default FALSE
 #' @return
-#' Returns and log messages produced by OTP, and will return the message "Graph built" if successful
+#' Returns and log messages produced by OTP, and will return the message
+#'     "Graph built" if successful
 #' @details
-#' The OTP .jar file can be downloaded from https://repo1.maven.org/maven2/org/opentripplanner/otp/
+#' The OTP .jar file can be downloaded from
+#'     https://repo1.maven.org/maven2/org/opentripplanner/otp/
 #'
 #' To build an OTP graph requires the following files to be in the directory
 #' specified by the dir variable.
 #'
 #' /graphs - A sub-directory
 #'
-#'   /default - A sub-directory with the name of the OTP router used in 'router' variable
+#'   /default - A sub-directory with the name of the OTP router used in router' variable
 #'
 #'     osm.pbf - Required, pbf file containing the Open Street Map
 #'
@@ -30,8 +36,10 @@
 #'
 #'     terrain.tif - Optional, GeoTiff image of terrain map
 #'
-#' The function will accept any file name for the .jar file, but it must be the only .jar file in that directory
-#' OTP can support multiple routers (e.g. different regions), each router must have its own sub-directory in the graphs directory
+#' The function will accept any file name for the .jar file, but it must be
+#' the only .jar file in that directory OTP can support multiple routers
+#' (e.g. different regions), each router must have its own sub-directory in
+#' the graphs directory
 #' @family setup
 #' @examples
 #' \dontrun{
@@ -47,11 +55,13 @@ otp_build_graph <- function(otp = NULL,
   # Run Checks
   checkmate::assert_numeric(memory, lower = 500)
   check <- otp_checks(otp = otp, dir = dir, router = router, graph = FALSE)
-  if(!check){
+  if (!check) {
     stop()
   }
-  message(paste0(Sys.time(),
-                 " Basic checks completed, building graph, this may take a few minutes"))
+  message(paste0(
+    Sys.time(),
+    " Basic checks completed, building graph, this may take a few minutes"
+  ))
 
   text <- paste0(
     "java -Xmx",
@@ -73,10 +83,10 @@ otp_build_graph <- function(otp = NULL,
 
   # Check for errors
   if (any(grepl("ERROR", set_up, ignore.case = TRUE)) & length(set_up) < 10) {
-    message(paste0(Sys.time()," Failed to build graph with message:"))
+    message(paste0(Sys.time(), " Failed to build graph with message:"))
[…]
