+++ tests/testthat/test_04_with_public_API.R
@@ -7,17 +7,18 @@ otpcon <- otp_connect(url = "https://api.digitransit.fi:443/routing/v1/routers/h
 context("Test the otp_plan function")
 
 test_that("basic routing", {
- route <- otp_plan(otpcon,
-   fromPlace = matrix(c(24.904353, 60.220485, 0, 0),
-                      ncol = 2,
-                      byrow = TRUE),
-   toPlace = c(24.952068, 60.192033),
-   fromID = c("A","Bug"),
-   toID = "B",
-   full_elevation = TRUE
- )
- expect_is(route, "sf")
- expect_true(nrow(route) == 1)
+  route <- otp_plan(otpcon,
+    fromPlace = matrix(c(24.904353, 60.220485, 0, 0),
+      ncol = 2,
+      byrow = TRUE
+    ),
+    toPlace = c(24.952068, 60.192033),
+    fromID = c("A", "Bug"),
+    toID = "B",
+    full_elevation = TRUE
+  )
+  expect_is(route, "sf")
+  expect_true(nrow(route) == 1)
 })
 
 
