### 2019-10-11 d404d89
Update vignettes
