+++ vignettes/prerequisites.Rmd
@@ -15,41 +15,35 @@ vignette: >
 This vignette provides guidance for people who new to OpenTripPlanner (OTP) and R.
 After you have R and OTP installed and have become familiar with them (e.g. by reading this vignette), we recommend working through [Getting Started vignette](https://itsleeds.github.io/opentripplanner/articles/opentripplanner.html).
 
+### Why would I want to use OpenTripPlanner for R?
+
+[OpenTripPlanner](http://www.opentripplanner.org/) (OTP) is a free, open-source, and cross-platform multi-modal route planner written in Java.  It is like having your own private version of Google Maps. OTP can be used to run public route-finding services such as https://ride.trimet.org but it can also be run on your own computer or server. If you want to analyse a transport networks OTP is a very useful tool. However, OTP is, as the name suggests, a Trip Planner, not analytical software.  So, while OTP can find the fastest route from A to B or find all the places that are within a 20-minute walk of C, it cannot answer a question like “how many people live within 10-minutes of a park?”  This is where R can help.  R can process multiple spatial datasets such as population densities and park locations but does not have a built-in journey planner. 
+
+This package allows you to use the trip planning power of OTP with the analytical might of R.
+
 ## What is this package for, and what is a package anyway?
 
 The `opentripplanner` R package makes it easier for R and OpenTripPlanner to communicate.
 Specifically, it allows you to do use R to control OTP and use it as a local routing service.
 For more on local versus remote routing services, see the Transportation chapter in [Geocomputation with R](https://geocompr.robinlovelace.net/transport.html).
 
-### What is OpenTripPlanner?
-
-OTP is a free, open-source, and cross-platform multi-modal route planner written in JAVA.  It is like having your own private version of Google Maps.
-
-Learn more about OTP at:
-
-* http://www.opentripplanner.org/
-* https://wiki.openstreetmap.org/wiki/OpenTripPlanner
-
-### What is R?
-
-R is a programming language and free software environment for statistical computing and graphics.  R has many capabilities for analysing data and writing software, but in this context, its ability to produce and analyse spatial data and maps is most relevant.
-
-You can find a lot more information about R at:
-
-* https://www.r-project.org/ 
-* https://en.wikipedia.org/wiki/R_(programming_language) 
-
-To get started with R, see An Introduction to R [online](https://cran.r-project.org/doc/manuals/r-release/R-intro.pdf) or from within R by running `help.start()`, or introductory tutorials such as DataCamp's [free](https://www.datacamp.com/courses/free-introduction-to-r) Introduction to R course.
-
 ### What is an R package?
 
 An R package is a small piece of software that extends the basic capabilities of R. It is a bit like how a new phone can do some things out of the box (make phone calls, send email) but you have to install apps to add extra abilities.
 
-### Why would I want to use OpenTripPlanner for R?
+### What are R and RStudio?
 
-OTP can be used to run public route finding services such as https://ride.trimet.org but it can also be run on your own private computer or server. If you want to analyse a transport network OTP is a very useful tool. However, OTP is, as the name suggests, a Trip Planner, not analytical software.  So, while OTP can find the fastest route from A to B or find all the places that are within a 20-minute walk of C, it cannot answer a question like “how many people live within 10-minutes of a park?”  This is where R can help.  R can process multiple spatial datasets such as population densities and park locations but does not have a built-in journey planner. 
[…]
