+++ vignettes/opentripplanner.Rmd
@@ -11,7 +11,7 @@ vignette: >
 
 ## Introduction
 
-This vignette is an introduction to OpenTripPlanner (OTP) - an open-source and cross-platform multi-modal route planner written in JAVA. It uses imported OpenStreetMap (OSM) data for routing on the street and path network and supports
+This vignette is an introduction to OpenTripPlanner (OTP) - an open-source and cross-platform multi-modal route planner written in Java. It uses imported OpenStreetMap (OSM) data for routing on the street and path network and supports
 multi-agency public transport routing through imported  [General Transit Feed Specification](https://developers.google.com/transit/gtfs/) (GTFS) feeds. It can also apply a digital elevation model to the OSM street network, allowing, for example, cycle-friendly routes to be requested. OTP has a web front-end that can be used by end-users and a sophisticated routing API.
 
 OTP works worldwide as long as there is OSM map coverage in the area. Support for transit timetables and elevation calculations are dependant on the available data. For more information on getting data see the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html).
@@ -176,7 +176,7 @@ Create a router called 'default', and include the GTFS and OSM files for the roa
          
 ```
 
-`router-config.json` is actually read when the OTP server is started (in the next section), but we add it now for convenience.
+`router-config.json` is actually read when the OTP server is started (in the next section), but we add it now for convenience. For more information on the config files see the [advanced features vignette](https://itsleeds.github.io/opentripplanner/articles/advanced_features.html).
 
 ### Getting Data for OTP
 
@@ -190,7 +190,7 @@ Note that OTP only really needs the road network from the OSM, but the Geofabrik
 
 #### Public Transport Timetables
 
-OTP can use public transport timetable data in the [GTFS format](https://developers.google.com/transit/gtfs/). You can find GTFS data for many regions at  [transitland](https://transit.land/).
+OTP can use public transport timetable data in the [GTFS format](https://developers.google.com/transit/gtfs/). You can find GTFS data for many regions at [transitland](https://transit.land/). For users in the UK see [UK2GTFS](https://itsleeds.github.io/UK2GTFS/)
 
 #### Elevation Data
 
