### 2019-10-11 dd856d0
New helper functions
