+++ vignettes/opentripplanner.Rmd
@@ -31,33 +31,33 @@ devtools::install_github("ITSLeeds/opentripplanner") # Requires the devtools pac
 library(opentripplanner)
 ```
 
-## Downloading OTP
+## Downloading OTP and the demonstation data
 
-Then you will need to download the OTP, for more details see the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html).
-The following code downloads the file into a temporary folder called `OTP` you may wish to change this to keep your results after closing R:
+We will build an example graph for the Isle of Wight using some example data provided for the package. A graph is what OTP uses to find routes, and must be built out of the raw data provided.
+
+### Creating a folder to store OTP and its data
+
+OTP expects its data to be stored in a specific structure, see [building your own otp graph](https://itsleeds.github.io/opentripplanner/articles/opentripplanner.html#building-your-own-otp-graph) for more details. We will make a folder called OTP in the temporary directory, you may wish to change this to keep your results after closing R.
 
 ```{r, eval=FALSE}
-# download file (may want to do this manually)
-dir.create(file.path(tempdir(), "OTP"))
+dir.create(file.path(tempdir(), "OTP")) 
 path_data <- file.path(tempdir(), "OTP")
-path_otp <- file.path(path_data, "otp.jar")
-u <- "https://repo1.maven.org/maven2/org/opentripplanner/otp/1.4.0/otp-1.4.0-shaded.jar"
-download.file(url = u, destfile = path_otp, mode = "wb")
 ```
 
-## Downloading Example Data
+### Downloading OTP
 
-We will build an example graph for the Isle of Wight using some example data provided for the package. A graph is what OTP uses to find routes, and must be built out of the raw data provided. First, we will download the sample data and put it in the folder structure that OTP expects.
+The `otp_dl_jar` function will download the OTP and save it in the folder we created. The function returns the path to the OTP jar file.
+
+```{r, eval=FALSE}
+path_otp <- otp_dl_jar(path_data)
+```
+
+### Downloading Example Data
+
+The `otp_dl_demo` function downloads the demonstation data and puts it in the correct structure.
 
 ```{r eval=FALSE}
-# Path to the sample data
-dir.create(file.path(path_data,"graphs")) # create a folder structure for the data
-dir.create(file.path(path_data,"graphs","default"))
-# Download example data
-download.file("https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip", 
-              destfile = file.path(path_data,"isle-of-wight-demo.zip") , mode="wb")
-unzip(file.path(path_data,"isle-of-wight-demo.zip"), exdir = file.path(path_data,"graphs","default"))
-unlink(file.path(path_data,"isle-of-wight-demo.zip"))
+otp_dl_demo(path_data)
 ```
 
 ## Building an OTP Graph
@@ -65,7 +65,6 @@ unlink(file.path(path_data,"isle-of-wight-demo.zip"))
 Now we can build the graph. This code will create a new file `Graph.obj` that will be saved in the location defined by `path_data`. 
 
 ```{r eval=FALSE}
-# Run the OTP and Build a graph object
 log <- otp_build_graph(otp = path_otp, dir = path_data) 
 ```
 
@@ -162,7 +161,7 @@ An OTP graph specifies every location in the region covered and how to travel be
 
 Our first task is to create the folder and file structure expected by OTP. This is a base directory called `otp` which contains a sub-directory called `graphs`. Directories created under `graphs` are known as OTP routers and contain all the files required to build a graph. A single OTP instance can host several routers, for example covering different regions. 
 
-Create a router called 'default', and include the GTFS and OSM files for the road network and default transit schedules for the railway on the Isle of Wight, along with `router-config.json`, as shown below:
[…]

[…]
