+++ R/otp-download.R
@@ -12,7 +12,7 @@
 #' @family setup
 #' @examples
 #' \dontrun{
-#'  otp_dl_jar(tempdir())
+#' otp_dl_jar(tempdir())
 #' }
 #' @export
 
@@ -20,12 +20,10 @@ otp_dl_jar <- function(path = NULL,
                        version = "1.4.0",
                        file_name = "otp.jar",
                        url = "https://repo1.maven.org/maven2/org/opentripplanner/otp") {
-
-  url <- paste0(url,"/",version,"/otp-",version,"-shaded.jar")
+  url <- paste0(url, "/", version, "/otp-", version, "-shaded.jar")
   destfile <- file.path(path, file_name)
   download.file(url = url, destfile = destfile, mode = "wb")
   return(destfile)
-
 }
 
 
@@ -39,27 +37,29 @@ otp_dl_jar <- function(path = NULL,
 #' @family setup
 #' @examples
 #' \dontrun{
-#'  otp_dl_demo(tempdir())
+#' otp_dl_demo(tempdir())
 #' }
 #' @export
 
 otp_dl_demo <- function(path_data = NULL,
-                       url = "https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip") {
-  if(!dir.exists(path_data)){
-    stop(paste0("Can't find folder ",path_data))
+                        url = "https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip") {
+  if (!dir.exists(path_data)) {
+    stop(paste0("Can't find folder ", path_data))
   }
-  if(!dir.exists(file.path(path_data,"graphs"))){
-    dir.create(file.path(path_data,"graphs"))
+  if (!dir.exists(file.path(path_data, "graphs"))) {
+    dir.create(file.path(path_data, "graphs"))
   }
-  if(!dir.exists(file.path(path_data,"graphs","default"))){
-    dir.create(file.path(path_data,"graphs","default"))
+  if (!dir.exists(file.path(path_data, "graphs", "default"))) {
+    dir.create(file.path(path_data, "graphs", "default"))
   }
 
-  download.file(url = url,
-                destfile = file.path(path_data,"isle-of-wight-demo.zip"),
-                mode="wb")
-  unzip(file.path(path_data,"isle-of-wight-demo.zip"),
-        exdir = file.path(path_data,"graphs","default"))
-  unlink(file.path(path_data,"isle-of-wight-demo.zip"))
-
+  download.file(
+    url = url,
+    destfile = file.path(path_data, "isle-of-wight-demo.zip"),
+    mode = "wb"
+  )
+  unzip(file.path(path_data, "isle-of-wight-demo.zip"),
+    exdir = file.path(path_data, "graphs", "default")
+  )
+  unlink(file.path(path_data, "isle-of-wight-demo.zip"))
 }
