+++ tests/testthat/test_01_without_OTP.R
@@ -179,13 +179,13 @@ test_that("otp_make_config tests", {
 })
 
 test_that("otp_build_graph input validation", {
-  expect_error(otp_build_graph(otp = paste0(path_opt,"Z"), dir = path_data),
+  expect_error(otp_build_graph(otp = paste0(path_otp,"Z"), dir = path_data),
     regexp = "File does not exist:"
   )
 })
 
 test_that("otp_setup input validation", {
-  expect_error(otp_setup(otp = paste0(path_opt,"Z"), dir = path_data),
+  expect_error(otp_setup(otp = paste0(path_otp,"Z"), dir = path_data),
     regexp = "File does not exist:"
   )
 })
