+++ README.Rmd
@@ -43,8 +43,13 @@ The package contains three groups of functions:
 
 Functions for setting up a local instance of OTP:
 
-1. `otp_build_graph()`  To make a OTP graph from raw data
-1. `otp_setup()`        To start up a local instance of OTP
+1. `otp_dl_jar()`          To download the OTP Jar file
+1. `otp_dl_demo()`         To download the demo data for the Isle of Wight
+1. `otp_build_graph()`     To make a OTP graph from raw data
+1. `otp_setup()`           To start up a local instance of OTP
+1. `otp_make_config()`     To make a config object
+1. `otp_validate_config()` To validate a config object
+1. `otp_write_config()`    To save a config object as a json file
 
 Functions for connecting to a local or remote instance of OTP:
 
