### 2019-10-11 9c1de81
fix bug in correct_distances
