### 2019-10-11 5b0df79
update vignettes
