+++ README.Rmd
@@ -30,11 +30,15 @@ To use OpenTripPlanner on your local computer you will need to install Java 8 an
 
 ### R Package
 
-Install the package with **devtools** as follows:
+Install the package using the **devtools** package as follows:
 
 ```{r install, eval=FALSE}
-install.packages("devtools") # If you do not already have the devtools package
+# If you do not already have the devtools package
+install.packages("devtools")
+# Install the package from GitHub
 devtools::install_github("ITSleeds/opentripplanner")
+# Load the package
+library(opentripplanner)
 ```
 
 ## Usage
@@ -58,7 +62,7 @@ Functions for connecting to a local or remote instance of OTP:
 Functions for retrieving data from OTP:
 
 1. `otp_plan()`         To get routes from A to B
-1. `otp_isochone()`     To get isochrone maps
+1. `otp_isochrone()`     To get isochrone maps
 1. `otp_geocode()`      To get the locations of named places e.g. road names
 
 Results are returned as [sf objects](https://cran.r-project.org/web/packages/sf/index.html)
