### 2019-10-23 e87e890
Increase code coverage
