### 2019-10-24 0717a76
no java on CI
