### 2019-10-31 #295 @mem48
Hi @geanders , just checking to see if there has been any progress on the review. I'm keen to get the package onto CRAN and I don't know the etiquette, should I wait for acceptance to ROpenSci or can I do the two processes in parallel?
