### 2019-11-08 #295 @geanders
@mem49: Since we lost one of the reviewers, I did the second review of this package, which I'm attaching here. I think that most of my comments and suggestions will be pretty quick to address (either small code changes or edits to documentation), so I'd recommend that you address those before submitting to CRAN, as I think you could do that pretty quickly.

- [x] **Vignette(s)** demonstrating major functionality that runs successfully locally **Note:** Most code is set to not evaluate, in case the Java OTP set-up has not been done on the system, but evaluating all the code locally worked for me after doing the Java OTP set-up.

>##### For packages co-submitting to JOSS
>
>- [x] The package has an **obvious research application** according to [JOSS's definition](http://joss.theoj.org/about#submission_requirements)
>
>The package contains a `paper.md` matching [JOSS's requirements](http://joss.theoj.org/about#paper_structure) with:
>
>- [x] **A short summary** describing the high-level functionality of the software
>- [x] **Authors:**  A list of authors with their affiliations
>- [x] **A statement of need** clearly stating problems the software is designed to solve and its target audience.
>- [x] **References:** with DOIs for all those that have one (e.g. papers, datasets, software).

   **Note:** I had some errors when I ran these locally (see below)

---

### Testing Error Messages

Here is the output from testing locally on my computer:

```
> devtools::check()
Updating opentripplanner documentation
Writing NAMESPACE
Loading opentripplanner
Writing NAMESPACE
── Building ─────────────────────────────────────────────────────────────────────────── opentripplanner ──
Setting env vars:
● CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
● CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
● CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
──────────────────────────────────────────────────────────────────────────────────────────────────────────
✔  checking for file ‘/Users/georgianaanderson/Documents/ropensci_review/opentripplanner-master/DESCRIPTION’ ...
─  preparing ‘opentripplanner’:
✔  creating vignettes (2.5s)
─  building ‘opentripplanner_0.1.0.0000.tar.gz’

── Checking ─────────────────────────────────────────────────────────────────────────── opentripplanner ──
Setting env vars:
● _R_CHECK_CRAN_INCOMING_USE_ASPELL_: TRUE
● _R_CHECK_CRAN_INCOMING_REMOTE_    : FALSE
● _R_CHECK_CRAN_INCOMING_           : FALSE
● _R_CHECK_FORCE_SUGGESTS_          : FALSE
── R CMD check ───────────────────────────────────────────────────────────────────────────────────────────
─  using log directory ‘/private/var/folders/83/ffl13dn505z0jvklj1y612kw0000gn/T/RtmpYlg7Lp/opentripplanner.Rcheck’
─  using R version 3.6.1 (2019-07-05)
─  using platform: x86_64-apple-darwin15.6.0 (64-bit)
─  using session charset: UTF-8
─  using options ‘--no-manual --as-cran’
✔  checking for file ‘opentripplanner/DESCRIPTION’
─  this is package ‘opentripplanner’ version ‘0.1.0.0000’
─  package encoding: UTF-8
✔  checking package namespace information ...
✔  checking package dependencies (869ms)
✔  checking if this is a source package
✔  checking if there is a namespace
✔  checking for executable files ...
✔  checking for hidden files and directories
✔  checking for portable file names ...
✔  checking for sufficient/correct file permissions
✔  checking serialization versions
✔  checking whether package ‘opentripplanner’ can be installed (1.3s)
✔  checking installed package size ...
✔  checking package directory ...
✔  checking for future file timestamps (1.4s)
✔  checking ‘build’ directory
N  checking top-level files ...
   Non-standard file/directory found at top level:
     ‘paper.bib’
✔  checking for left-over files ...
✔  checking index information ...
✔  checking package subdirectories ...
✔  checking R files for non-ASCII characters ...
✔  checking R files for syntax errors ...
✔  checking whether the package can be loaded ...
✔  checking whether the package can be loaded with stated dependencies ...
✔  checking whether the package can be unloaded cleanly ...
✔  checking whether the namespace can be loaded with stated dependencies ...
✔  checking whether the namespace can be unloaded cleanly ...
✔  checking dependencies in R code (508ms)
✔  checking S3 generic/method consistency (347ms)
✔  checking replacement functions ...
✔  checking foreign function calls ...
N  checking R code for possible problems (2.2s)
   Warning: <anonymous>: ... may be used in an incorrect context

   otp_get_results: ... may be used in an incorrect context
✔  checking Rd files ...
✔  checking Rd metadata ...
N  checking Rd line widths ...
   Rd file 'otp_dl_demo.Rd':
     \usage lines wider than 90 characters:
          url = "https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip")

   These lines will be truncated in the PDF manual.
✔  checking Rd cross-references ...
✔  checking for missing documentation entries ...
✔  checking for code/documentation mismatches ...
✔  checking Rd \usage sections (466ms)
✔  checking Rd contents ...
✔  checking for unstated dependencies in examples ...
✔  checking installed files from ‘inst/doc’ ...
✔  checking files in ‘vignettes’ ...
✔  checking examples (677ms)
✔  checking for unstated dependencies in ‘tests’ ...
─  checking tests ...
E  Running ‘testthat.R’ [2s/11s] (10.7s)
     ── 2. Failure: can get lsoa points (@test_03_with_OTP.R#14)  ─────────────────────────────────────────────
     nrow(lsoa) == 89 isn't true.

     ══ testthat results  ═════════════════════════════════════════════════════════════════════════════════════
     [ OK: 68 | SKIPPED: 17 | WARNINGS: 100 | FAILED: 2 ]
     1. Failure: test otp_checks without graph, with files (@test_02_internal_funcs.R#134)
     2. Failure: can get lsoa points (@test_03_with_OTP.R#14)

     Error: testthat unit tests failed
     In addition: Warning messages:
     1: In for (pkg in package) { :
       closing unused connection 5 (<-localhost:11066)
     2: In for (pkg in package) { :
       closing unused connection 4 (<-localhost:11066)
✔  checking for unstated dependencies in vignettes ...
✔  checking package vignettes in ‘inst/doc’ ...
✔  checking re-building of vignette outputs (2.8s)

   See
     ‘/private/var/folders/83/ffl13dn505z0jvklj1y612kw0000gn/T/RtmpYlg7Lp/opentripplanner.Rcheck/00check.log’

── R CMD check results ─────────────────────────────────────────────────── opentripplanner 0.1.0.0000 ────
Duration: 24.8s

❯ checking tests ...
  See below...

❯ checking top-level files ... NOTE
  Non-standard file/directory found at top level:
    ‘paper.bib’

❯ checking R code for possible problems ... NOTE
  Warning: <anonymous>: ... may be used in an incorrect context

  otp_get_results: ... may be used in an incorrect context

❯ checking Rd line widths ... NOTE
  Rd file 'otp_dl_demo.Rd':
    \usage lines wider than 90 characters:
         url = "https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip")

  These lines will be truncated in the PDF manual.

── Test failures ─────────────────────────────────────────────────────────────────────────── testthat ────

> library(testthat)
> library(opentripplanner)
>
> test_check("opentripplanner")
trying URL 'https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip'
Content type 'application/octet-stream' length 8309996 bytes (7.9 MB)
==================================================
downloaded 7.9 MB

trying URL 'https://repo1.maven.org/maven2/org/opentripplanner/otp/1.4.0/otp-1.4.0-shaded.jar'
Content type 'application/java-archive' length 61838786 bytes (59.0 MB)
==================================================
downloaded 59.0 MB

── 1. Failure: test otp_checks without graph, with files (@test_02_internal_funcs.R#134)  ────────────────
otp_checks(...) isn't false.

── 2. Failure: can get lsoa points (@test_03_with_OTP.R#14)  ─────────────────────────────────────────────
nrow(lsoa) == 89 isn't true.

══ testthat results  ═════════════════════════════════════════════════════════════════════════════════════
[ OK: 68 | SKIPPED: 17 | WARNINGS: 100 | FAILED: 2 ]
1. Failure: test otp_checks without graph, with files (@test_02_internal_funcs.R#134)
2. Failure: can get lsoa points (@test_03_with_OTP.R#14)

Error: testthat unit tests failed
In addition: Warning messages:
1: In for (pkg in package) { :
  closing unused connection 5 (<-localhost:11066)
2: In for (pkg in package) { :
  closing unused connection 4 (<-localhost:11066)

1 error ✖ | 0 warnings ✔ | 3 notes ✖
```

It looks like the main concern is with getting the "lsoa" points file in (which I also
had problems with when running example code---see note below on that point). Also, any
idea on why 17 of the tests skipped? Is this a concern? Finally, the tests only ran
correctly if I started in a fresh R session. Maybe it would be useful to add some
code to the testing to clean-up any temporary directory and the environment before
running the tests to make the process a bit more robust (this isn't a major point, so
no worries if there's not an easy way to do this)?

Oh my goodness, what a very cool package this is! I really enjoyed playing around with it.

While it does require an initial investment from the user in terms of set-up, the updates to
the documentation put that in the range of many R users, so I think it has the potential to
reach a broad audience (and I don't see any obvious way to make the set-up process easier).

I identified some typos or areas in the documentation where different wording might be clearer.
I will submit a separate pull request for the package code to provide some suggestions for that.

I had the following specific comments, questions, and suggestions:

- For functions that save a file to the user's computer, it would be nice to add a line to
documentation clarifying that something will be written to the user's computer, as well
as perhaps a message printed when the function runs letting the user know a file is being
written and where. I think the functions doing this are: otp_build_graph, otp_dl_demo,
otp_dl_jar, and otp_write_config. With the nice documentation, I don't think this should
be a shock to users that these functions are writing files to their computer, but I still
think it's nice to make that very clear to users.
- For the help files, when `router` is an option (e.g., `otp_build_graph`), there is a note
that it "must match with contents of dir". Could you clarify this point? I wasn't sure how
to interpret it. Also, for functions with this option, it would be nice in the help file to
include a link to more documentation on OTP and what routers are for OTP (and why you might
want to use something other than the default).
- For the help file for `otp_build_graph`, what is an "analyst feature"? Again, perhaps a
link to the OTP documentation on this point would be helpful in this helpfile.
- In the helpfile for `otp_build_graph`, it mentions in the Return Value section that this
function returns and logs messages. Is it just printing out the returned messages, or does
it write them in a file somewhere? The "log" term made me think it's writing these to a file
somewhere, but then it wasn't clear to me where in the function code it was writing those
messages (if it was). Perhaps change "log" to "print" here if the messages are just being
printed out directly at the R console.
- I think it's wonderful that you have the example in the vignette set up so that everything's
being downloaded to a temporary directory, so the user doesn't have a lot of clean-up to do
from running the example. Also, you make it very clear that the user might want to use a
non-temporary directory for a real project. I think it would be helpful to users if you could
explain this point a bit more (perhaps with some example code somewhere of what you'd do for the
set-up for a "real" project, where you want to be able to work on it more or reproduce the
analysis later, rather than having everything delete when the R session is closed). For example,
in one of the examples, you're referring to an OTP JAR file that's located in "C:/otp/otp.jar".
Would this be a better place to save the OTP JAR when running `otp_dl_jar` if you think you'll
want to use this program a lot? Otherwise, the user's re-downloading the JAR file each time they
do an analysis, which doesn't seem necessary (have I interpreted what's going on here correctly?).
This is a change that could wait for later versions (in terms of adding a bit to the vignette
documentation), but I think at some point would be very helpful, as I imagine some users right
now will end up using the example code with the temporary directory set-up every time they use
this package, which would work but be a bit inefficient.
- For the `otp_connect` function, perhaps it would be helpful to paste the status code from
`check_router` into the error message if that code is not 200? Maybe this extra information
would help the user diagnose problems with connecting if they have them?
- For the `otp_isochrone` function, there is a note in the helpfile that this feature is known
to not work correctly with any mode other than TRANSIT. If this is the case, perhaps it would
be good to edit the assertion code in this function so that an error, or a least a message,
is returned if the user tries to use any of the other modes? Also, this seems worth mentioning in
the vignette section on Isochrones, as well as in the helpfile, as I imagine with the nice
vignettes, many users will mostly learn about this package from the vignettes rather than the
helpfiles. A warning only mentioned in the helpfile might not make it through to all users.
- For functions that use `maxWalkDistance` and `minTransferTime`, please add into the helpfiles
the units that should be used for the numbers given for these parameters. Also, for values
like these passed directly through to the Java OTP session, could you add links to the OTP
help documentation so that a user could get a better idea how to use these options?
- For functions that use the `mode` parameter, can a vector of lengths longer than one
be used to allow for multiple modes of transport? If not, could you clarify by change to
something like "length-one character vector" rather than just "character vector" in the
parameter definition in the helpfile? This might be another point where a link to the OTP
documentation (directly in the parameter definition) would help users figure out how to use
this parameter.
- There's a note in the helpfile for `otp_plan` about a bug related to the elevation
profile only being returned for the first leg of the route. Is this still the case? From the
vignette, it looks like you're able to get some elevation information throughout the route.
Could you revisit the wording on this part of the helpfile and make sure it's still the case?
- In any of the examples that start with `otp_connect`, perhaps it would also make sense to
include `opt_stop()` at the end of the example code? And in general, what are the repercussions
if a user forgets to run `opt_stop`? I guess that the process will still be shut down the next
time the person restarts their computer? It would be helpful to know if it is a very big deal
if someone forgets to stop the Java process.
- What would happen if a user had two (or more) different versions of Java installed? Since
OTP requires a very specific version of Java, this seems possible. In one of the checking functions,
you have R running a system call to "java -version" and then checking the version number
installed based on the output, so I'm curious what would happen with this piece of the code if
the user has versions 1.8 and a different version installed.
- Really nice use of assertions throughout the code!
- The vignettes are just really, really great! Some of the parts that I found particularly
nice include details about the memory and file sizes that users can expect for different parts of
the process and connections with the OTP software's documentation. The examples in the
"Advanced" vignette are also very nice and I think will attract many users to try out the package.
- For the `otp_stop` function, there's a typo in the message returned by the function.
- The naming convention for the functions (all start with `otp_`) is really thoughtful.
- Somewhere in the vignettes, it would be helpful to add a table with the name of all key
package functions in one column and a brief description in the next.
- For the longer-term with this package, since it does require some set-up, it might be interesting
to consider creating a container (e.g., with Docker) that someone could download with all the
set-up incorporated (would that work?).
- It would be nice to address (briefly) reproducibility. It looks like the package defaults to
always pull the same version of OTP, so perhaps mention that and that, as a result, if a user
reruns code a while later, they should be using the same version of OTP?
- In the "Advanced Features" vignette, it might be helpful to clarify the language about
error messages when you're referring to error messages from the Java OTP session that are
transmitted by R. When I read "error messages", I assumed they were R error messages, so
it was confusing why they wouldn't kill the process, but I'm pretty sure that these are
instead error or warning messages from the Java process that pass through R and are printed out
by R.
- What is an "OD matrix"? Origin-destination? Perhaps spell this out the first time you refer
to it (unless it would be very clear to all transport researchers).
- For the "Advanced Features" vignette, you may want to consider showing what the first few lines
of the created objects would look like, when you can. For example, it would be interesting to see
what part of the `routes_matrix` looks like in the vignette.
- The "Known Issues" vignette is a very nice addition!
- For the examples in the help functions, I think it's reasonable that they're "Not run", as
otherwise I think there will be some problems on CRAN, since it won't have the proper
installation of the OTP JAR.
- Progress messages / bars are really nice
- For "fromPlace" and "toPlace", must the locations be given with a certain number of
significant digits for the latitude and longitude? If so, please clarify in the help files.
- I got the following warning when I tried to read the centroids data directly from GitHub
(but had no problem when I downloaded and read the file from a local location):

```
> lsoa <- sf::st_read("https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/centroids.gpkg",
+                     stringsAsFactors = FALSE)
Reading layer `GeoJSONSeq' from data source `https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/centroids.gpkg' using driver `GeoJSONSeq'
Simple feature collection with 0 features and 0 fields
bbox:           xmin: NA ymin: NA xmax: NA ymax: NA
epsg (SRID):    4326
proj4string:    +proj=longlat +datum=WGS84 +no_defs
There were 50 or more warnings (use warnings() to see the first 50)
```

- Once of the vignettes uses `spread` from `tidyr`. This function is now deprecated---could you
update to use `pivot_wider`?
