### 2019-11-08 05cdc10
adjust tests
move tests to dedicated folders
