+++ R/otp-connect.R
@@ -6,7 +6,7 @@
 #'
 #' @param hostname A string, e.g. "ec2-34-217-73-26.us-west-2.compute.amazonaws.com".
 #'     Optional, default is "localhost".
-#' @param router A string, e.g. "UK2018". Optional, default is "default".
+#' @param router A string, e.g. "UK2018". Optional, default is "default". OTP can support multiple routers see advanced vignette for details.
 #' @param url If a nonstandard URL stucture is used provide a full url,
 #'     default is NULL
 #' @param port A positive integer. Optional, default is 8080.
@@ -82,7 +82,7 @@ otp_connect <- function(hostname = "localhost",
       message("Router ", make_url(otpcon), " exists")
       return(otpcon)
     } else {
-      stop("Router ", make_url(otpcon), " does not exist")
+      stop("Router ", make_url(otpcon), " does not exist. Error code ", check_router(otpcon))
     }
   } else {
     return(otpcon)
