### 2019-11-08 5770e08
clarify multiple modes
