+++ R/otp-isochrone.R
@@ -3,8 +3,10 @@
 #' @param otpcon OTP connection object produced by otp_connect()
 #' @param fromPlace Numeric vector, Longitude/Latitude pair,
 #'     e.g. `c(-0.134649, 51.529258,)`
-#' @param mode Character vector of modes of travel valid values
-#'     TRANSIT, WALK, BICYCLE, CAR, BUS, RAIL, default CAR
+#' @param mode character vector of one or more modes of travel valid values
+#'     TRANSIT, WALK, BICYCLE, CAR, BUS, RAIL, default CAR. Not all
+#'     combinations are valid e.g. c("WALK","BUS") is valid but
+#'     c("WALK","CAR") is not.
 #' @param date_time POSIXct, a date and time, defaults to current
 #'     date and time
 #' @param arriveBy Logical, Whether the trip should depart or
