+++ tests/testthat/test_02_internal_funcs.R
@@ -113,20 +113,23 @@ skip_otp <- function() {
 context("Test otp_checks")
 dir.create(file.path(tempdir(), "otp2"))
 path_data <- file.path(tempdir(), "otp2")
-path_otp <- file.path(tempdir(), "otp2","otp.jar")
+path_otp <- file.path(tempdir(), "otp2", "otp.jar")
 
 
 test_that("test otp_checks without graph, missing files", {
-  expect_error(otp_checks(otp = path_otp,
-                                            dir = path_data,
-                                            router = "default",
-                                            graph = FALSE),
-               regexp = "does not exist")
+  expect_error(otp_checks(
+    otp = path_otp,
+    dir = path_data,
+    router = "default",
+    graph = FALSE
+  ),
+  regexp = "does not exist"
+  )
 })
 
 
 dir.create(file.path(path_data, "graphs"))
-dir.create(file.path(path_data, "graphs","default"))
+dir.create(file.path(path_data, "graphs", "default"))
 file.create(path_otp)
 
 # test_that("test otp_checks without graph, with files", {
@@ -138,14 +141,17 @@ file.create(path_otp)
 # })
 
 test_that("test otp_checks with graph, missing files", {
-  expect_error(otp_checks(otp = path_otp,
-                                            dir = path_data,
-                                            router = "default",
-                                            graph = TRUE),
-               regexp = "File does not exist")
+  expect_error(otp_checks(
+    otp = path_otp,
+    dir = path_data,
+    router = "default",
+    graph = TRUE
+  ),
+  regexp = "File does not exist"
+  )
 })
 
-file.create(file.path(path_data, "graphs","default","Graph.obj"))
+file.create(file.path(path_data, "graphs", "default", "Graph.obj"))
 
 # test_that("test otp_checks with graph, with files", {
 #   skip_otp()
