### 2019-11-08 c2e6f19
update manual
