### 2019-11-08 dbe6fee
update vignettes
