### 2019-11-08 #295 @mem48
@geanders lots to go through, but I think I've fixed most of these.

> It looks like the main concern is with getting the "lsoa" points file in (which I also had problems with when running example code---see note below on that point). Also, any idea on why 17 of the tests skipped? Is this a concern? Finally, the tests only ran correctly if I started in a fresh R session

I fixed some badly designed tests, the skips are because you don't have Sys.getenv("I_have_OTP") see https://github.com/ITSLeeds/opentripplanner/blob/master/README.md Its a simple toggle to disable tests that require the correct version of Java.

>     For functions that save a file to the user's computer, it would be nice to add a line to
>     documentation clarifying that something will be written to the user's computer.

Warning message added giving file path

>     For the help files, when router is an option (e.g., otp_build_graph), there is a note
>     that it "must match with contents of dir". Could you clarify this point? I wasn't sure how
>     to interpret it. Also, for functions with this option, it would be nice in the help file to
>     include a link to more documentation on OTP and what routers are for OTP (and why you might
>     want to use something other than the default).

Updated the vignette with explanation.

>     For the help file for otp_build_graph, what is an "analyst feature"? Again, perhaps a
>     link to the OTP documentation on this point would be helpful in this helpfile.

Updated the vignette with explanation.

>     In the helpfile for otp_build_graph, it mentions in the Return Value section that this
>     function returns and logs messages. Is it just printing out the returned messages, or does
>     it write them in a file somewhere? The "log" term made me think it's writing these to a file
>     somewhere, but then it wasn't clear to me where in the function code it was writing those
>     messages (if it was). Perhaps change "log" to "print" here if the messages are just being
>     printed out directly at the R console.

Updated the help with explanation. There is no log file, just printed to console.

>     I think it's wonderful that you have the example in the vignette set up so that everything's
>     being downloaded to a temporary directory, so the user doesn't have a lot of clean-up to do
>     from running the example. Also, you make it very clear that the user might want to use a
>     non-temporary directory for a real project. I think it would be helpful to users if you could
>     explain this point a bit more (perhaps with some example code somewhere of what you'd do for   the
>     set-up for a "real" project, where you want to be able to work on it more or reproduce the
>     analysis later, rather than having everything delete when the R session is closed). For example,
>     in one of the examples, you're referring to an OTP JAR file that's located in "C:/otp/otp.jar".
>     Would this be a better place to save the OTP JAR when running otp_dl_jar if you think you'll
>     want to use this program a lot? Otherwise, the user's re-downloading the JAR file each time they
>     do an analysis, which doesn't seem necessary (have I interpreted what's going on here correctly?).

Yes you are right, Updated the vignette with explination.

>     For the otp_connect function, perhaps it would be helpful to paste the status code from
>     check_router into the error message if that code is not 200?

Done

>     For the otp_isochrone function, there is a note in the helpfile that this feature is known to not work correctly with any mode other than TRANSIT.

It is, but this is something that will be fixed by the main OTP team. I think giving extra warnings based on the version of OTP is a bit beyond the remit of the R package. The R package is doing its job correctly and OTP provides their own documentation which overs how OTP itself works.

>     For functions that use maxWalkDistance and minTransferTime, please add into the helpfiles the units that should be used for the numbers given for these parameters.

Done

> Also, for values like these passed directly through to the Java OTP session, could you add links to the OTP help documentation so that a user could get a better idea how to use these options?

They are rewriting their documentation at the moment, so perhaps best left until later?

>     For functions that use the mode parameter, can a vector of lengths longer than one be used to allow for multiple modes of transport?

Help updated

>     There's a note in the helpfile for otp_plan about a bug related to the elevation profile only being returned for the first leg of the route. Is this still the case?

It is, the vignette only shows a single leg journey, the bug occurs when you get a walk, transit, walk route. You only get elevation data for the first walk leg.

>     In any of the examples that start with otp_connect, perhaps it would also make sense to include opt_stop() at the end of the example code?

It is not necessary to stop OPT, in most cases, it stops automatically when RStudio is closed. The function is mostly there for testing, allowing users to force OTP to stop if needed.

What would happen if a user had two (or more) different versions of Java installed?

With multiple versions of java there is still a "default" one. I haven't tested this extensively as users who can install multiple version of java are probably expert users.

>     Somewhere in the vignettes, it would be helpful to add a table with the name of all key
package functions in one column and a brief description in the next.

Do you mean https://itsleeds.github.io/opentripplanner/reference/index.html ?

>     For the longer-term with this package, since it does require some set-up, it might be interesting to consider creating a container (e.g., with Docker)

Good idea, but then there would need to be a Docker tutorial. I think about it for the long term.

>     It would be nice to address (briefly) reproducibility. It looks like the package defaults to always pull the same version of OTP, so perhaps mention that and that, as a result, if a user reruns code a while later, they should be using the same version of OTP?

otp_dl_jar() has a url argument that selects the version of OTP. I had assumed that once new versions come out I would update the url once I've checked that the package is working well with the new version.

>     In the "Advanced Features" vignette, it might be helpful to clarify the language about error messages when you're referring to error messages from the Java OTP session that are transmitted by R.

Updated vignette

>     What is an "OD matrix"?

Updated vignette

>     For the "Advanced Features" vignette, you may want to consider showing what the first few lines
of the created objects would look like, when you can.

To do that I will have to actually set up a working instance of OTP when building the vignettes, which is hard to make work on the servers which don't usually have Java. I could put in some screenshots if you think it is important.

>     For the examples in the help functions, I think it's reasonable that they're "Not run", as otherwise, I think there will be some problems on CRAN, since it won't have the proper installation of the OTP JAR.

All the example should have the \dontrun tag, can you give a specific example?

>     For "fromPlace" and "toPlace", must the locations be given with a certain number of significant digits for the latitude and longitude?

It doesn't matter, lest digits is just less precise.

>     I got the following warning when I tried to read the centroids data directly from GitHub (but had no problem when I downloaded and read the file from a local location):

I can't reproduce, could your firewall be blocking R from downloading files?
