### 2019-11-13 b7f55ad
Add warning about isochones
