+++ R/otp-download.R
@@ -1,10 +1,10 @@
 #' Download OTP Jar File
 #'
 #' @description
-#' Download the OTP jar file from mavern.org
+#' Download the OTP jar file from maven.org
 #'
 #' @param path path to folder where OTP is to be stored
-#' @param version a character string ov version number default is "1.4.0"
+#' @param version a character string of the version number default is "1.4.0"
 #' @param file_name file name to give the otp default "otp.jar"
 #' @param url URL to the dowload server
 #' @return
