+++ man/otp_stop.Rd
@@ -29,10 +29,13 @@ otp_stop(kill_all = FALSE)
 }
 }
 \seealso{
-Other setup: \code{\link{otp_build_graph}},
-  \code{\link{otp_dl_demo}}, \code{\link{otp_dl_jar}},
-  \code{\link{otp_make_config}}, \code{\link{otp_setup}},
-  \code{\link{otp_validate_config}},
-  \code{\link{otp_write_config}}
+Other setup: 
+\code{\link{otp_build_graph}()},
+\code{\link{otp_dl_demo}()},
+\code{\link{otp_dl_jar}()},
+\code{\link{otp_make_config}()},
+\code{\link{otp_setup}()},
+\code{\link{otp_validate_config}()},
+\code{\link{otp_write_config}()}
 }
 \concept{setup}
