+++ README.md
@@ -1,43 +1,23 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
+ITSleeds <a href='https://itsleeds.github.io/'><img src='man/figures/logo.png' align="right" height=215/></a>
+=============================================================================================================
 
-# ITSleeds <a href='https://itsleeds.github.io/'><img src='man/figures/its.jpg' align="right" height=215/></a>
+[![Travis build status](https://travis-ci.org/ITSLeeds/opentripplanner.svg?branch=master)](https://travis-ci.org/ITSLeeds/opentripplanner) [![Coverage status](https://codecov.io/gh/ITSLeeds/opentripplanner/branch/master/graph/badge.svg)](https://codecov.io/github/ITSLeeds/opentripplanner?branch=master) [![Build status](https://ci.appveyor.com/api/projects/status/gqp3smc04as3qg85?svg=true)](https://ci.appveyor.com/project/layik/opentripplanner-05ana) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 
-[![Travis build
-status](https://travis-ci.org/ITSLeeds/opentripplanner.svg?branch=master)](https://travis-ci.org/ITSLeeds/opentripplanner)
-[![Coverage
-status](https://codecov.io/gh/ITSLeeds/opentripplanner/branch/master/graph/badge.svg)](https://codecov.io/github/ITSLeeds/opentripplanner?branch=master)
-[![Build
-status](https://ci.appveyor.com/api/projects/status/gqp3smc04as3qg85?svg=true)](https://ci.appveyor.com/project/layik/opentripplanner-05ana)
-[![Project Status: Active – The project has reached a stable, usable
-state and is being actively
-developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+OpenTripPlanner for R
+=====================
 
-# OpenTripPlanner for R
+The goal of OpenTripPlanner for R is to provide a simple R interface to [OpenTripPlanner (OTP)](https://www.opentripplanner.org/). The OTP is a multimodal trip planning service written in Java. For more information on what OTP is, see the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html).
 
-The goal of OpenTripPlanner for R is to provide a simple R interface to
-[OpenTripPlanner (OTP)](https://www.opentripplanner.org/). The OTP is a
-multimodal trip planning service written in Java. For more information
-on what OTP is, see the [prerequisites
-vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html).
+This package can be used to interface with a remote instance of OTP (e.g. a website) or help you set up and manage a local version of OTP for private use. Basic setup and routing functions are outlined in the [getting started vignette](https://itsleeds.github.io/opentripplanner/articles/opentripplanner.html), while advanced functionality such as batch routing, isochrones, and customised setup is described in the [advanced features vignette](https://itsleeds.github.io/opentripplanner/articles/advanced_features.html).
 
-This package can be used to interface with a remote instance of OTP
-(e.g. a website) or help you set up and manage a local version of OTP
-for private use. Basic setup and routing functions are outlined in the
-[getting started
-vignette](https://itsleeds.github.io/opentripplanner/articles/opentripplanner.html),
-while advanced functionality such as batch routing, isochrones, and
-customised setup is described in the [advanced features
-vignette](https://itsleeds.github.io/opentripplanner/articles/advanced_features.html).
-
-## Installation
+Installation
+------------
 
 ### OpenTripPlanner
 
[…]
