### 2019-11-19 #295 @mem48
@geanders I've finished spell checking the documentation. I think the package is good to go unless you want any other changes?
