### 2019-11-19 #295 @kimnewzealand
There are few suggestions to follow up on, please let me know if you had a chance to review these:

> Depending on rOpenSci's current view on Datacamp, perhaps remove the tutorial references
> Add a reference hyperlink for the .pbf format
> Add a reference hyperlink for the GeoTIFF format
> In this line Note that OTP only really needs the road network from the OSM. So for large areas, you may wish to remove unneeded data from the file using osmconvert and osmfilter and remove the and and explain where (OSM or R?) you can remove this unneeded data.
> Move this section from getting started to prerequisites as it feels like it part of the initial set up, and the user can work through this set up once before then installing or loading the R package.
> Before you can use the package, you should set-up the OpenTripPlanner Java program. You will need to download the latest version of OTP from https://repo1.maven.org/maven2/org/opentripplanner/otp/ , at time of writing the latest version of OTP is version 1.3 so you would want to get https://repo1.maven.org/maven2/org/opentripplanner/otp/1.3.0/otp-1.3.0-shaded.jar
>
> Include a mention of the memory parameter or include another line of code in case the user encounters memory issues in this step:
> log <- otp_build_graph(otp = path_otp, dir = path_data, memory = 1)
> There are few places ion the README and vignettes where the . is missing at the end of the sentences
