### 2019-11-19 #295 @kimnewzealand
I have also  updated the checkmark for the Community guidelines given:
`Bug reports and comments are welcome as Github Issues and code submissions as Pull Requests.`
