### 2019-11-23 #295 @mem48
Hi @kimnewzealand

> Add a reference hyperlink for the .pbf format
> Add a reference hyperlink for the GeoTIFF format
> In this line Note that OTP only really needs the road network from the OSM. .....
> There are few places ion the README and vignettes where the . is missing at the end of the sentences

These are done.

The getting data and the bits about downloading the .jar file have been rewritten too. In particular I've added a new functions `otp_dl_jar()` and  `otp_dl_demo()` which simplify these stages.

I would prefer to keep the section on "getting your own data" in the getting started vignette, because I think most users would want to go through the vignettes with the demo data first then come back to using their own data.

If finding your own data is made part of the prerequisites, then new users may think that they need their own data to complete the rest of the tutorials.

> Include a mention of the memory parameter or include another line of code in case the user encounters memory issues in this step:
> log <- otp_build_graph(otp = path_otp, dir = path_data, memory = 1)

Memory use is now explained in the hardware section of the prerequisites vignette and repeated in the getting started vignette.

I've read the readme and added punctuation as required.
