+++ tests/testthat/test_03_with_OTP.R
@@ -8,7 +8,7 @@ skip_no_otp <- function() {
 
 context("Test the download of the LSOA file")
 
-lsoa <- sf::read_sf("https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/centroids.gpkg")
+lsoa <- sf::read_sf("https://github.com/ropensci/opentripplanner/releases/download/0.1/centroids.gpkg")
 test_that("can get lsoa points", {
   expect_is(lsoa, "sf")
   expect_true(nrow(lsoa) == 89)
