+++ R/otp-download.R
@@ -43,7 +43,7 @@ otp_dl_jar <- function(path = NULL,
 #' @export
 
 otp_dl_demo <- function(path_data = NULL,
-                        url = "https://github.com/ITSLeeds/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip") {
+                        url = "https://github.com/ropensci/opentripplanner/releases/download/0.1/isle-of-wight-demo.zip") {
   if (!dir.exists(path_data)) {
     stop(paste0("Can't find folder ", path_data))
   }
