+++ README.md
@@ -1,20 +1,41 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-OpenTripPlanner for R <a href='https://itsleeds.github.io/'><img src='man/figures/logo.png' align="right" height=180/></a>
-==========================================================================================================================
 
-[![Travis build status](https://travis-ci.org/ITSLeeds/opentripplanner.svg?branch=master)](https://travis-ci.org/ITSLeeds/opentripplanner) [![Coverage status](https://codecov.io/gh/ITSLeeds/opentripplanner/branch/master/graph/badge.svg)](https://codecov.io/github/ITSLeeds/opentripplanner?branch=master) [![Build status](https://ci.appveyor.com/api/projects/status/gqp3smc04as3qg85?svg=true)](https://ci.appveyor.com/project/layik/opentripplanner-05ana) [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+# OpenTripPlanner for R <a href='https://itsleeds.github.io/'><img src='man/figures/logo.png' align="right" height=180/></a>
 
-The goal of OpenTripPlanner for R is to provide a simple R interface to [OpenTripPlanner (OTP)](https://www.opentripplanner.org/). The OTP is a multimodal trip planning service written in Java. For more information on what OTP is, see the [prerequisites vignette](https://itsleeds.github.io/opentripplanner/articles/prerequisites.html).
+[![Travis build
+status](https://travis-ci.org/ropensci/opentripplanner.svg?branch=master)](https://travis-ci.org/ropensci/opentripplanner)
+[![Coverage
+status](https://codecov.io/gh/ropensci/opentripplanner/branch/master/graph/badge.svg)](https://codecov.io/github/ropensci/opentripplanner?branch=master)
+[![Build
+status](https://ci.appveyor.com/api/projects/status/gqp3smc04as3qg85?svg=true)](https://ci.appveyor.com/project/ropensci/opentripplanner-05ana)
+[![Project Status: Active – The project has reached a stable, usable
+state and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
 
-This package can be used to interface with a remote instance of OTP (e.g. a website) or help you set up and manage a local version of OTP for private use. Basic setup and routing functions are outlined in the [getting started vignette](https://itsleeds.github.io/opentripplanner/articles/opentripplanner.html), while advanced functionality such as batch routing, isochrones, and customised setup is described in the [advanced features vignette](https://itsleeds.github.io/opentripplanner/articles/advanced_features.html).
+The goal of OpenTripPlanner for R is to provide a simple R interface to
+[OpenTripPlanner (OTP)](https://www.opentripplanner.org/). The OTP is a
+multimodal trip planning service written in Java. For more information
+on what OTP is, see the [prerequisites
+vignette](https://docs.ropensci.org/opentripplanner/articles/prerequisites.html).
 
-Installation
-------------
+This package can be used to interface with a remote instance of OTP
+(e.g. a website) or help you set up and manage a local version of OTP
+for private use. Basic setup and routing functions are outlined in the
+[getting started
+vignette](https://docs.ropensci.org/opentripplanner/articles/opentripplanner.html),
+while advanced functionality such as batch routing, isochrones, and
+customised setup is described in the [advanced features
+vignette](https://docs.ropensci.org/opentripplanner/articles/advanced_features.html).
+
+## Installation
 
 ### OpenTripPlanner
 
[…]
