+++ DESCRIPTION
@@ -16,10 +16,11 @@ Description: An R interface for OpenTripPlanner <http://www.opentripplanner.org/
     OpenTripPlanner (OTP) is an open source platform for multi-modal and multi-agency
     journey planning written in Java. OTP provides a REST API which OpenTripPlanner
     for R uses to connect to OTP.
+ This package has been peer-reviewed by rOpenSci (v. 0.1.0.0).
 Language: EN-GB
 License: file LICENSE
-URL: https://github.com/ITSLeeds/opentripplanner, https://itsleeds.github.io/opentripplanner/
-BugReports: https://github.com/ITSLeeds/opentripplanner/issues
+URL: https://github.com/ropensci/opentripplanner, https://docs.ropensci.org/opentripplanner/
+BugReports: https://github.com/ropensci/opentripplanner/issues
 Encoding: UTF-8
 LazyData: true
 Imports:
@@ -31,7 +32,7 @@ Imports:
     jsonlite,
     pbapply,
     sf
-RoxygenNote: 7.0.0
+RoxygenNote: 7.0.1
 Suggests: 
     covr,
     knitr,

[…]
