+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: opentripplanner
 Title: OpenTripPlanner for R
-Version: 0.1.0.0000
+Version: 0.2.0.0000
 Authors@R: c(
     person("Malcolm", "Morgan", email = "m.morgan1@leeds.ac.uk", role = c("aut","cre"), 
       comment = c(ORCID = "0000-0002-9488-9183")),
@@ -16,7 +16,7 @@ Description: An R interface for OpenTripPlanner <http://www.opentripplanner.org/
     OpenTripPlanner (OTP) is an open source platform for multi-modal and multi-agency
     journey planning written in Java. OTP provides a REST API which OpenTripPlanner
     for R uses to connect to OTP.
- This package has been peer-reviewed by rOpenSci (v. 0.1.0.0).
+ This package has been peer-reviewed by rOpenSci (v. 0.2.0.0).
 Language: EN-GB
 License: file LICENSE
 URL: https://github.com/ropensci/opentripplanner, https://docs.ropensci.org/opentripplanner/
