+++ .travis.yml
@@ -14,10 +14,10 @@ addons:
       - libgeos-dev
       - libproj-dev
 
-before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
+#before_deploy: Rscript -e 'remotes::install_cran("pkgdown")'
 after_success:
     - Rscript -e 'covr::codecov()'
-deploy:
-  provider: script
-  script: Rscript -e 'pkgdown::deploy_site_github()'
-  skip_cleanup: true
+#deploy:
+#  provider: script
+#  script: Rscript -e 'pkgdown::deploy_site_github()'
+#  skip_cleanup: true
