### 2019-11-29 #295 @mem48
Hi @stefaniebutland  I'd be happy to write a short introduction to the package. Is it ok to rework some of the vignettes or does it have to completly original content?
