### 2019-11-29 #295 @stefaniebutland
Hi @mem48. It does have to be original content. For something short, you could consider writing a [tech note](https://ropensci.org/technotes/), "Package updates, technical know-how, and small announcements related to rOpenSci software development".

fyi @Robinlovelace is writing a big-picture post about stplanr for publication in December. Noting in case that influences your content.
