### 2019-11-29 #295 @Robinlovelace
Yes this should go in there, all these things join up and a big picture overview is just what's needed right now. More soon, xmas deadline noted! By the way huge congratulations Malcolm!
