### 2019-11-29 1628775
Reduce n. chars in docs, fix #34
