### 2019-11-30 a8a4252
extra cran checks
