+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: opentripplanner
 Title: OpenTripPlanner for R
-Version: 0.2.0.0001
+Version: 0.2.0.2
 Authors@R: c(
     person("Malcolm", "Morgan", email = "m.morgan1@leeds.ac.uk", role = c("aut","cre"), 
       comment = c(ORCID = "0000-0002-9488-9183")),
