+++ paper.bib
@@ -9,13 +9,14 @@
   month = {jan},
   year = {2018},
   pages = {245--254},
-  note = {Last visited on 2019-04-17TZ},
+  note = {Last visited on 04/17/2019},
 }
 
 @inproceedings{bast_fast_2010,
   series = {Lecture {Notes} in {Computer} {Science}},
   title = {Fast {Routing} in {Very} {Large} {Public} {Transportation} {Networks} {Using} {Transfer} {Patterns}},
-  booktitle = {Algorithms – {ESA} 2010},
+  doi = {10.1007/978-3-642-15775-2_25},
+  booktitle = {Algorithms ? {ESA} 2010},
   publisher = {Springer Berlin Heidelberg},
   author = {Hannah Bast and Erik Carlsson and Arno Eigenwillig and Robert Geisberger and Chris Harrelson and Veselin Raychev and Fabien Viger},
   editor = {Mark {de Berg} and Ulrich Meyer},
