+++ paper.bib
@@ -16,7 +16,7 @@
   series = {Lecture {Notes} in {Computer} {Science}},
   title = {Fast {Routing} in {Very} {Large} {Public} {Transportation} {Networks} {Using} {Transfer} {Patterns}},
   doi = {10.1007/978-3-642-15775-2_25},
-  booktitle = {Algorithms ? {ESA} 2010},
+  booktitle = {Algorithms - {ESA} 2010},
   publisher = {Springer Berlin Heidelberg},
   author = {Hannah Bast and Erik Carlsson and Arno Eigenwillig and Robert Geisberger and Chris Harrelson and Veselin Raychev and Fabien Viger},
   editor = {Mark {de Berg} and Ulrich Meyer},
