### 2019-12-02 aafa920
Add JOSS badge
