### 2019-12-02 73d45c9
tweak for CRAN
matrix are becoming matrix arrays
