### 2019-12-10 e98588e
cran fixes
