+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: opentripplanner
-Title: Setup and connect to OpenTripPlanner
+Title: Setup and connect to 'OpenTripPlanner'
 Version: 0.2.0.3
 Authors@R: c(
     person("Malcolm", "Morgan", email = "m.morgan1@leeds.ac.uk", role = c("aut","cre"), 
@@ -12,9 +12,9 @@ Authors@R: c(
       comment = c(ORCID = "0000-0003-1912-4890"))
     )
 Maintainer: Malcolm Morgan <m.morgan1@leeds.ac.uk>
-Description: Setup and connect to OpenTripPlanner (OTP) <http://www.opentripplanner.org/>.
+Description: Setup and connect to 'OpenTripPlanner' (OTP) <http://www.opentripplanner.org/>.
     OTP is an open source platform for multi-modal and multi-agency
-    journey planning written in Java. The package allows you to manage a local version or 
+    journey planning written in 'Java'. The package allows you to manage a local version or 
     connect to remote OTP server e.g. <https://api.digitransit.fi:443/routing/v1/routers/hsl>.
     This package has been peer-reviewed by rOpenSci (v. 0.2.0.0).
 Language: EN-GB
