### 2020-01-03 4c302b9
fix bug in tests where file path is wrong
