+++ DESCRIPTION
@@ -32,7 +32,7 @@ Imports:
     jsonlite,
     pbapply,
     sf
-RoxygenNote: 7.0.1
+RoxygenNote: 7.0.2
 Suggests: 
     covr,
     knitr,
