### 2020-01-04 bfb56fa
ad linux java note
