### 2020-01-06 #295 @maelle
Closing this, which doesn't prevent the convo about blog posts from continuing. :wink:
