### 2020-01-06 #295 @stefaniebutland
> have been prioritising academic projects and holiday
wisdom!

> One question where I imagine rOpenSci people can help is release on CRAN.
rOpenSci Slack is a great place to ask questions like this. Lots of CRAN experience and specific helpful answers

When you're ready, email me for blog post follow up

Cheers
